% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_permission.R
\name{ask_permission}
\alias{ask_permission}
\title{Ask permission to write settings on disk}
\usage{
ask_permission()
}
\value{
'logical' if TRUE, the user authorised saving in this directory.
}
\description{
Ask users for permission to create a .sen2r settings directory,
in which to store the files paths.json and apihub.txt.
}
\note{
License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2019) \email{lbusett@gmail.com}

Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
