% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_example_param_file.R
\name{build_example_param_file}
\alias{build_example_param_file}
\title{Build an example JSON parameter file}
\usage{
build_example_param_file(
  json_path = tempfile(fileext = "_sen2r_params.json"),
  overwrite = TRUE
)
}
\arguments{
\item{json_path}{Path of the output file. Default is to save it on a
temporary file, whose path is returned.}

\item{overwrite}{Logical value: should existing output file be
overwritten? (default: TRUE)}
}
\value{
The path of the created file.
}
\description{
Function used to write JSON parameter file.
A function is provided instead than a json file to ensure
directories to match the user folder tree.
}
\note{
License: GPL 3.0
}
\examples{
build_example_param_file()
}
\references{
L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020).
"sen2r": An R toolbox for automatically downloading and preprocessing
Sentinel-2 satellite data. \emph{Computers & Geosciences}, 139, 104473. DOI:
\href{https://doi.org/10.1016/j.cageo.2020.104473}{10.1016/j.cageo.2020.104473},
URL: \url{http://sen2r.ranghetti.info/}.
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
