% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_creds.R
\name{check_api_creds}
\alias{check_api_creds}
\title{check_api_creds}
\usage{
check_api_creds(user, pwd)
}
\arguments{
\item{user}{\code{string} User Name}

\item{pwd}{\code{string}  password}
}
\value{
TRUE if credentials are valid, FALSE elsewhere
}
\description{
Checks ifprovided username and password are valid Scihub credentials
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
if(interactive()){
 check_api_creds("user", "user")
 }
}
}
\author{
Lorenzo Busetto, phD (2019) \email{busetto.l@irea.cnr.it}
}
