% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies_gui.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check package dependencies}
\usage{
check_dependencies()
}
\description{
The function allows to graphically check that all the
dependencies are installed.
}
\details{
This package needs some external dependencies to run:
\itemize{
\item Python
\item GDAL
\item Wget
\item sen2cor
\item s2download
}

This function opens a GUI which allows to check that these dependencies
are installed. This check is highly suggested before using the library for
the fist time, in order to avoid errors.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{

check_dependencies()
}
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
