% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_param_list.R
\name{check_param_list}
\alias{check_param_list}
\title{Check a fidolasen parameter list}
\usage{
check_param_list(pm, type = "string")
}
\arguments{
\item{type}{Type of the output (see \link{print_message} for details).}
}
\value{
Depending on \code{type} argument, output can be a vector of errors
(if \code{type = "string"}), the first error occurred (if \code{type = "error"})
or a set of warnings (if \code{type = "warning"}). If no errors occur,
output is NULL.
}
\description{
Check that the parameter list (or JSON parameter file)
is in the correct format, and then speficied values are coherent with
parameters.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
