% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_s2_paths_legacy.R
\name{compute_s2_paths_legacy}
\alias{compute_s2_paths_legacy}
\title{Compute names of S2 file to be generated (old function)}
\usage{
compute_s2_paths_legacy(pm, s2_list_l1c, s2_list_l2a, paths, list_prods,
  out_ext, force_tiles = FALSE, check_tmp = TRUE, ignorelist)
}
\arguments{
\item{pm}{List of input parameters.}

\item{s2_list_l1c}{Names and paths of input SAFE level-1C products.}

\item{s2_list_l2a}{Names and paths of input SAFE level-2A products.}

\item{paths}{Named vector of required paths. It must contain elements
named \code{"out"}, \code{"indices"}, \code{"tiles"}, \code{"merged"} and \code{"warped"}.}

\item{list_prods}{Character vector with the values of the
products to be processed (accepted values: "TOA", "BOA", "SCL", "TCI").}

\item{out_ext}{Extension (character) of output products.}

\item{force_tiles}{(optional) Logical: passed to \link{safe_shortname} (default: FALSE).}

\item{check_tmp}{(optional) Logical: if TRUE (default), temporary files
are also searched when \code{_exi} names are computed;
if FALSE, only non temporary files are searched.}

\item{ignorelist}{Vector of output files to be ignored.}

\item{index_ext}{Extension (character) of index products.}

\item{tiles_ext}{Extension (character) of tiled products.}

\item{merged_ext}{Extension (character) of merged products.}

\item{warped_ext}{Extension (character) of warped products.}

\item{rgb_ext}{Extension (character) of RGB products.}

\item{sr_masked_ext}{Extension (character) of masked products of SR products.}
}
\value{
A list, containing the following vectors:
\code{tiles_names_exp}, \code{merged_names_exp}, \code{warped_names_exp},
\code{masked_names_exp}, \code{out_names_exp}, \code{indices_names_exp},
\code{indices_names_new}, \code{out_names_req}, \code{out_names_new},
\code{masked_names_new}, \code{warped_names_req}, \code{warped_names_reqforrgb},
\code{warped_names_new},  \code{merged_names_req}, \code{merged_names_reqforrgb},
\code{merged_names_new}, \code{tiles_names_req}, \code{tiles_names_new},
\code{safe_names_l1c_req}, \code{safe_names_l2a_req}.
}
\description{
\code{compute_s2_paths} is an internal function
(to be used within \code{\link[=sen2r]{sen2r()}})
which computes the names of the required output image files
(see details).
The function was splitted from \code{\link[=sen2r]{sen2r()}} because this code
is called twice (and to shorten the main function).
This function is deprecated, \link{compute_s2_paths} is used instead.
}
\details{
\code{compute_s2_paths_legacy} was structured in the following way:
\enumerate{
\item Retrieve the file names expected to be present at the
end of the processing chain (suffix \code{_exp}):
\itemize{
\item \code{tiles_names_}
\item \code{merged_names_}
\item \code{warped_names_}
\item \code{masked_names_}
\item \code{out_names_} (= \code{warped_names_} or \code{subset(warped_names_)+masked_names_})
\item \code{indices_names_}
}
\item Compute the file names expected to be created
(suffixes \code{_req} and \code{_new}, see below)
(this operation is done in reverse order).

Meaning of the suffixes \code{_exp}, \code{_req} and \code{_new}
(here and for all the script):
\itemize{
\item \code{_exp}: full names of the files expected to be present at the
end of the processing chain (already existing or not);
\item \code{_req}: names of the files required for the next step
(e.g. tiles_names_req are required to perform s2_merge())
\item \code{_new}: names of the required files not existing yet (expected
to be created).
}
}

With \code{overwrite=TRUE}, all these vectors are equal
(e.g. \code{merged_names_exp = merged_names_req = merged_names_new}),
because all is overwritten.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
