% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_l2agipp.R
\name{copy_l2agipp}
\alias{copy_l2agipp}
\title{Copy L2A_GIPP.xml in sen2r}
\usage{
copy_l2agipp()
}
\value{
TRUE if the file was copied, FALSE elsewhere (invisible output)
}
\description{
Internal function to copy L2A_GIPP.xml from default Sen2Cor
directory to sen2r. After that, user will allow editing Sen2Cor options
in sen2r without affecting standalone Sen2Cor behaviour.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
copy_l2agipp()
}
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
