% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_s2_dop.R
\name{create_s2_dop}
\alias{create_s2_dop}
\title{Create the database of S2 orbits and doy of passage}
\usage{
create_s2_dop(json_path = NA, force = FALSE)
}
\arguments{
\item{json_path}{(optional) The path of the output JSON file.
\emph{Warning}: to create a file which will be usable by the package,
this option must be left to NA (default location is within the
package installation). Edit this only to create the file in another
place for external use.}

\item{force}{(optional) Logical: if FALSE (default), the db is created only
if missing or not updated; if TRUE, it is created in any case.}
}
\value{
The path of the json file
}
\description{
The internal function build a database with the base DOY of
passage across each Sentinel-2A orbit (which is used in function \code{s2_dop}).
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
