% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dontuse.R
\name{dontuse}
\alias{dontuse}
\title{Don't use}
\usage{
dontuse()
}
\description{
Don't use this function.
}
\details{
This function simply call an \code{rgdal} function, so to be able to
include package \code{rgdal} among imported dependencies.
This is necessary since using \code{raster} without having \code{rgdal} installed
causes troublings in R 4.
}
