% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidolasen_s2.R
\name{fidolasen_s2}
\alias{fidolasen_s2}
\title{Find, download and preprocess Sentinel-2 images}
\usage{
fidolasen_s2(param_list = NULL, sel_sensor = NA, online = NA,
  overwrite_safe = NA, rm_safe = NA, step_atmcorr = NA,
  steps_reqout = NA, timewindow = NA, timeperiod = NA, extent = NA,
  s2tiles_selected = NA, list_prods = NA, reference_path = NA, res = NA,
  unit = NA, proj = NA, resampling = NA, outformat = NA,
  compression = NA, overwrite = NA, path_l1c = NA, path_l2a = NA,
  path_tiles = NA, path_out = NA, path_tiles_subdirs = NA,
  path_out_subdirs = NA)
}
\arguments{
\item{param_list}{List of input parameters: it can be both an
R list or the path of a JSON file.
If some parameters are passed both as elements of \code{param_list}
and as function arguments, the values passed as function
arguments are considered.
If some parameters are missing in \code{param_list} and are not
provided as arguments, default values will be used.
Use the function \code{\link[=s2_gui]{s2_gui()}} to create a complete list of
parameters.}

\item{sel_sensor}{=NA,}

\item{online}{=NA,}

\item{overwrite_safe}{=NA,}

\item{rm_safe}{=NA,}

\item{step_atmcorr}{=NA,}

\item{steps_reqout}{=NA,}

\item{timewindow}{=NA,}

\item{timeperiod}{=NA,}

\item{extent}{=NA,}

\item{s2tiles_selected}{=NA,}

\item{list_prods}{=NA,}

\item{reference_path}{=NA,}

\item{res}{=NA,}

\item{unit}{=NA,}

\item{proj}{=NA,}

\item{resampling}{=NA,}

\item{outformat}{=NA,}

\item{compression}{=NA,}

\item{overwrite}{=NA,}

\item{path_l1c}{=NA,}

\item{path_l2a}{=NA,}

\item{path_tiles}{=NA,}

\item{path_out}{=NA,}

\item{path_tiles_subdirs}{=NA,}

\item{path_out_subdirs}{=NA}

\item{timewindow}{=NA,}
}
\description{
The function is a wrapper to perform the entire
processing chain to find, download and pre-process Sentinel-2
data. Input is a set of parameters that can be passed with a
list or file (parameter \code{param_list}) or singularly (see the
descriptions of all the other parameters).
}
