% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs2nc_getElements.R
\name{fs2nc_getElements}
\alias{fs2nc_getElements}
\title{Get information from S2 short name}
\usage{
fs2nc_getElements(s2_name)
}
\arguments{
\item{s2_name}{A Sentinel-2 product name in the fidolasen-S2
naming convention.}
}
\value{
A list of the output metadata.
}
\description{
This accessory function extracts metadata included in
the name of a Sentinel-2 product which follows the fidolasen-S2
naming convention (see \link{s2_shortname}).
}
\note{
License: GPL 3.0
}
\examples{
# Define product name
fs2nc_examplename <-
  "/path/of/the/product/S2A1C_20170603_022_32TQQ_TOA_20.tif"

# Return metadata
fs2nc_getElements(fs2nc_examplename)
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
