% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_warp.R
\name{gdal_warp}
\alias{gdal_warp}
\title{Clip, reproject and warp raster files}
\usage{
gdal_warp(srcfiles, dstfiles, of = NULL, ref = NULL, tr = NULL,
  te = NULL, t_srs = NULL, r = NULL, dstnodata = NULL, ...)
}
\arguments{
\item{srcfiles}{A vector of input file paths (managed by GDAL).}

\item{dstfiles}{A vector of corresponding output file paths.}

\item{of}{The output format (use the short format name). Default is
the format of every input filename.}

\item{ref}{Path of the raster taken as reference: if provided, other
parameters regarding the output grid (\code{te}, \code{tr}, \code{t_srs}) are ignored,
and the output files are built with the same grid (alignment and
resolution) of \code{ref} raster.}

\item{tr}{Numeric. (\code{c(xres,yres)}). set output file resolution
(in target georeferenced units).}

\item{te}{Georeferenced extents of output file to be created
(xmin,ymin,xmax,ymax)}

\item{t_srs}{A vector of corresponding output file paths.}

\item{r}{A vector of corresponding output file paths.}

\item{dstnodata}{A vector of corresponding output file paths.}

\item{...}{Additional parameters of \link{gdalwarp} or \link{gdal_translate}
(different from \code{s_srs}, \code{t_srs}, \code{te}, \code{tr}, \code{ts} and \code{of}).}
}
\description{
The function applies \link{gdal_translate} or \link{gdalwarp}
to clip, reproject and/or warp raster files. The choice of the
algorythm is based on the comparison between input and output
projections (\link{gdal_translate} if they are equal, \link{gdalwarp} elsewhere).
If not specified, the output format of each file is the same of the
corresponding source file.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
ex_sel <- c("/path/of/existing/input/file.tif",
            "/path/of/existing/input/anotherfile.jp2")
ex_ref <- "/path/of/the/reference/file.jp2"
ex_out <- c("/path/of/the/output/file.tif",
            "/path/of/the/output/anotherfile.jp2")

gdal_warp(ex_sel, ex_out, ex_ref, dstnodata=0, overwrite=TRUE)
}
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
