% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalpython_utils.R
\name{gdalpython_utils}
\alias{gdalpython_utils}
\title{Interface to GDAL Python-based utilities}
\usage{
gdalpython_utils(
  util,
  source,
  destination,
  formula = character(0),
  options = character(0),
  quiet = FALSE
)
}
\arguments{
\item{util}{Character: one among \code{"calc"} and \code{"fillnodata"}
(other utilities were not implemented, since they are not used by sen2r).}

\item{source}{path of input layer(s); for \code{calc} this can be more than one.}

\item{destination}{Path of the output layer.}

\item{formula}{(for \code{util = "calc"}) Calculation in gdalnumeric syntax using
\code{+}, \code{-}, \code{/}, \code{*}, or any \code{numpy} array functions (i.e. \code{log10()}).}

\item{options}{Character vector with GDAL options.}

\item{quiet}{Logical: if TRUE, suppress printing of output for info
(this argument is ignored in case package \code{sys} is not installed).}
}
\value{
A logical (invisible) indicating success (i.e., TRUE);
in case of failure, an error is raised.
}
\description{
This accessory function interfaces with GDAL
Python-based utilities (sen2r must be interfaced with a runtime GDAL
environment, see \code{check_gcal()}).
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
