% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatype.R
\name{get_rastype}
\alias{get_rastype}
\title{Check in the input is a \code{Raster} object or raster file}
\usage{
get_rastype(in_rast, abort = TRUE)
}
\arguments{
\item{in_rast}{name of an \code{R} object, or \code{character} giving the full path
to a spatial file}

\item{abort}{If TRUE, and \code{in_rast} is neither a raster object or
filename, send an error message and abort, Default: TRUE}
}
\value{
\code{character} equal to "rastfile" (if \code{in_rast} is a raster file),
\code{rastobject} (if \code{in_rast} is a \code{R} raster object) or \code{NA} if it is
neither (unless \code{abort} == TRUE)
}
\description{
Check if a \code{R} object or a filename correspond to a valid \code{Raster}
object, to a raster file or none of the above. Useful to detect which kind
of input is passed to a function and abort / do something in the case of
"wrong" input.
}
\note{
Functions \link{get_spatype}, \link{get_rastype} and \link{get_vectype} come from
package \href{http://lbusett.github.io/sprawl}{sprawl}.
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
