% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatype.R
\name{get_vectype}
\alias{get_vectype}
\title{check the "spatial type" of an object or file}
\usage{
get_vectype(in_vect, abort = TRUE)
}
\arguments{
\item{in_vect}{name of an \code{R} object, or \code{character} giving the full path
to a vector file}

\item{abort}{If TRUE, and \code{in_vect} is neither a spatial object or
filename, send an error message and abort, Default: TRUE}
}
\value{
\code{character} equal to "vectfile" (if \code{in_vect} is a raster file),
\code{spobject} (if \code{in_vect} is of class \code{spatial}), \code{sfobject} (if \code{in_vect}
is of class \code{sf}), or \code{NA} if it is neither (unless \code{abort} == TRUE)
}
\description{
Check if a \code{R} object or a filename correspond to a valid vector
object, to a vector file or none of the above. Useful to detect which kind
of input is passed to a function and abort / do something in the case of
"wrong" input.
}
\note{
Functions \link{get_spatype}, \link{get_rastype} and \link{get_vectype} come from
package \href{http://lbusett.github.io/sprawl}{sprawl}.
}
