% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gipp.R
\name{read_gipp}
\alias{read_gipp}
\alias{set_gipp}
\alias{reset_gipp}
\title{Manage GIPP parameters for Sen2Cor}
\usage{
read_gipp(gipp_names, gipp_path = NA)

set_gipp(gipp = list(), gipp_path = NA, use_dem = NA)

reset_gipp(gipp_path = NA)
}
\arguments{
\item{gipp_names}{Character vector with the names of the parameters
to be read.}

\item{gipp_path}{Character path of the output GIPP XML file.
If NA (default), the default sen2r GIPP path is used.}

\item{gipp}{(optional) Ground Image Processing Parameters (GIPP)
(see \code{\link[=set_gipp]{set_gipp()}} for further details).
Elements whose name is missing in the XML file are skipped.}

\item{use_dem}{Logical, determining if a DEM should be set for being used
for topographic correction in the XML specified with argument \code{gipp_path}
(see \code{\link[=sen2cor]{sen2cor()}} for further details).}
}
\value{
\code{\link[=read_gipp]{read_gipp()}} returns a named list of GIPP with the required parameters
(values not found in the XML are skipped).

\code{\link[=set_gipp]{set_gipp()}} and \code{\link[=reset_gipp]{reset_gipp()}} return NULL
(functions are called for their side effects).
}
\description{
\code{\link[=read_gipp]{read_gipp()}} reads Ground Image Processing Parameters (GIPP)
from an XML file.

\code{\link[=set_gipp]{set_gipp()}} modifies values of a list of GIPP in an XML file.

\code{\link[=reset_gipp]{reset_gipp()}} restores default GIPP values (using values
set in the \code{L2A_GIPP.xml} default Sen2Cor file).
}
\note{
License: GPL 3.0
}
\examples{
\donttest{
# Read default values
read_gipp(c("dem_directory", "dem_reference"))
# Edit one value and save the output as temporary file
set_gipp(list(DEM_Directory = NA), gipp_path = gipp_temp <- tempfile())
# Read the parameters in the created temporary files
read_gipp(c("DEM_Directory", "DEM_Reference"), gipp_path = gipp_temp)
# Read the original (not edited) XML file
read_gipp(c("DEM_Directory", "DEM_Reference"))
# Reset to default Sen2Cor GIPP values
reset_gipp(gipp_path = gipp_temp)
# Read again values
read_gipp(c("DEM_Directory", "DEM_Reference"), gipp_path = gipp_temp)
}
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
