% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_s2download.R
\name{import_s2download}
\alias{import_s2download}
\title{Import s2download python module}
\usage{
import_s2download(...)
}
\arguments{
\item{...}{optional parameters of import() function}
}
\value{
s2download python module
}
\description{
s2download is a collection of python scripts used to download
and correct Sentinel-2 images, and it is required by RSPrePro.
This internal function check if they are installed and imports them.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
