% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_python.R
\name{init_python}
\alias{init_python}
\title{Initialise python}
\usage{
init_python()
}
\value{
NULL (the function is called for its side effects)
}
\description{
Internal function to set the environmental variables
required to run Python-based GDAL utilities on Windows.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
