% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_aria2.R
\name{install_aria2}
\alias{install_aria2}
\title{Download and install aria2.}
\usage{
install_aria2(aria2_dir = system.file(package = "sen2r"), force = FALSE)
}
\arguments{
\item{aria2_dir}{(optional) Path where aria2 executable will be installed
(default: the package path).}

\item{force}{(optional) Logical: if TRUE, install even if it is already
installed (default is FALSE).}
}
\value{
The path of aria2
}
\description{
This function download and install standalone version of
\href{https://aria2.github.io}{aria2} for Windows.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
install_aria2()
}
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
