% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sen2cor.R
\name{install_sen2cor}
\alias{install_sen2cor}
\alias{link_sen2cor}
\title{Download and install (or link) Sen2Cor}
\usage{
install_sen2cor(
  sen2cor_dir,
  version = "2.5.5",
  use_dem = NA,
  gipp = NULL,
  force = FALSE
)

link_sen2cor(sen2cor_dir, use_dem = NA, gipp = NULL)
}
\arguments{
\item{sen2cor_dir}{Path where sen2cor is being installed or searched.}

\item{version}{(optional) Character: Sen2Cor version (one among
'2.5.5' - default - and '2.8.0').}

\item{use_dem}{(optional) Logical:
if TRUE, Sen2Cor is set to use a Digital Elevation Model for topographic
correction (reflecting what is done for Level-2A SAFE images provided by ESA Hub);
if FALSE, it is set not to perform topographic correction (reflecting the
current default Sen2Cor behaviour);
if NA (default), the option set in the user's \code{L2A_GIPP.xml} Sen2Cor
configuration file is respected (in case the user never edited it,
the current default setting is not to perform topographic correction).

\emph{Notes}:
\enumerate{
\item this argument simply set the \code{DEM_Directory} parameter of the default
sen2r GIPP XML file to \verb{~/.sen2r/srtm90} or \code{NONE} in case
\code{use_dem} is respectively TRUE or FALSE.
For further customisation, set argument \code{gipp} properly.
\item Currently the default value is NA in order to grant backward
compatibility. In a future release of sen2r, the default value will be
set to TRUE, so to grant homogeneity between Level-2A products downloaded
from ESA Hub and generated using Sen2Cor.
\item If both \code{use_dem} is defined and \code{DEM_Directory} in argument \code{gipp} is
specified, the second choice overrides the first one.
}}

\item{gipp}{(optional) List of Ground Image Processing Parameters (GIPP)
to be modified after installing Sen2Cor with respect to default values
(see \href{[set_gipp]}{set_gipp()} for detail about using this argument).
Default behaviour is to maintain default values (with the exception of
parameter \code{DEM_Directory}, whih is managed by argument \code{use_dem}).}

\item{force}{(optional) Logical: if TRUE, install even if it is already
installed (default is FALSE).}
}
\value{
NULL (the function is called for its side effects)
}
\description{
\code{\link[=install_sen2cor]{install_sen2cor()}} downloads and installs standalone version of
\href{http://step.esa.int/main/third-party-plugins-2/sen2cor}{Sen2Cor}.

\code{link_sen2cor()} links an existing standalone version of
\href{http://step.esa.int/main/third-party-plugins-2/sen2cor}{Sen2Cor} to sen2r.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
install_sen2cor(sen2cor_dir = tempdir(), use_dem = TRUE)
# ( use a non-temporary folder path instead of tempdir() )
}
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
