% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_wget.R
\name{install_wget}
\alias{install_wget}
\title{Download and install wget.}
\usage{
install_wget(wget_dir = system.file(package = "sen2r"), force = FALSE)
}
\arguments{
\item{wget_dir}{(optional) Path where wget executable will be installed
(default: the package path).}

\item{force}{(optional) Logical: if TRUE, install even if it is already
installed (default is FALSE).}
}
\value{
The path of wget
}
\description{
This function download and install standalone version of
\href{https://eternallybored.org/misc/wget}{GNU Wget} for Windows.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
install_wget()
}
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
