% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_extent.R
\name{load_extent_bbox}
\alias{load_extent_bbox}
\alias{load_extent_vectfile}
\alias{load_extent_draw}
\title{Insert an extent}
\usage{
load_extent_bbox()

load_extent_vectfile()

load_extent_draw(extent_ns_name)
}
\arguments{
\item{extent_ns_name}{Name of the namespace to be used}
}
\description{
Internal functions and modal dialogs to specify an extent
in the GUI.
}
\note{
License: GPL 3.0
}
\references{
L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020).
"sen2r": An R toolbox for automatically downloading and preprocessing
Sentinel-2 satellite data. \emph{Computers & Geosciences}, 139, 104473. DOI:
\href{https://doi.org/10.1016/j.cageo.2020.104473}{10.1016/j.cageo.2020.104473},
URL: \url{http://sen2r.ranghetti.info/}.
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
