% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskapply_parallel.R
\name{maskapply_parallel}
\alias{maskapply_parallel}
\title{Parallel masking over SR products}
\usage{
maskapply_parallel(in_rast, in_mask, outpath, tmpdir = NA, binpaths,
  NAflag = 0, parallel = TRUE, minrows = NULL, datatype = "INT2S",
  overwrite = FALSE)
}
\arguments{
\item{in_rast}{Input raster Stack.}

\item{in_mask}{Input Raster to be used as mask (pixesl with values 1 are not
masked, pixels with values 0 are masked).}

\item{outpath}{Path of the output file.}

\item{tmpdir}{(optional) Path where intermediate VRT will be created.
Default is in a temporary directory.}

\item{binpaths}{list of paths of binaries.}

\item{NAflag}{(optional) NA value to be used where mask is applied.}

\item{parallel}{(optional) Logical: if TRUE, masking is conducted using parallel
processing, to speed-up the computation for large rasters.
The number of cores is automatically determined; specifying it is also
possible (e.g. \code{parallel = 4}).
If FALSE (default), single core processing is used.}

\item{minrows}{(optional) Argument to be passed to \link[raster:blockSize]{raster::blockSize}.}

\item{datatype}{(optional) data type of the output raster.}

\item{overwrite}{(optional) Logical value: should existing output files be
overwritten? (default: FALSE)}

\item{format}{(optional) Format of the output file (in a
format recognised by GDAL). Default is the same format of input images
(or "GTiff" in case of VRT input images).}
}
\description{
Internal function to apply cloud mask to SR products.
It is used inside \code{s2_mask} only on multiband virtual products
(because if the output is a physical file there is no gain in using
this procedure).
}
\note{
License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2018) \email{lbusett@gmail.com}

Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
