% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_path.R
\name{normalize_path}
\alias{normalize_path}
\title{Express file paths in canonical Form depending on the operating system}
\usage{
normalize_path(path, ...)
}
\arguments{
\item{path}{character vector of file paths}

\item{...}{additional parameters passed to \link{normalizePath} (i.e. \code{mustWork}).}
}
\value{
The paths normalized.
}
\description{
Accessory function wrapper for \code{normalizePath()} in Linux
and \code{shortPathName(normalizePath())} in Windows.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
