% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_check.R
\name{path_check}
\alias{path_check}
\title{Check a path}
\usage{
path_check(path, mustbe_empty = FALSE, mustbe_writable = TRUE)
}
\arguments{
\item{path}{\code{string} full path to a folder}

\item{mustbe_empty}{\code{logical} if TRUE, accept only empty directories}

\item{mustbe_writable}{\code{logical} if TRUE, accept only directories with write
permissions}
}
\description{
Accessory functions to check that a directory exists and
the user have write permissions on it (to be used in a Shiny context)
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}

Lorenzo Busetto, PhD (2019)
}
\keyword{internal}
