% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_message.R
\name{print_message}
\alias{print_message}
\title{print_message}
\usage{
print_message(..., type, sep = "", date = FALSE, date_format = "")
}
\arguments{
\item{...}{\code{R} objects which are concatenated.}

\item{type}{\code{character} Type of the output (accepted values: 'message'
for a diagnostic message, 'string' for a character output, 'cat' for the
output of \code{\link[base:cat]{base::cat()}} function, 'error' and 'warning' for an error or
warning message. Intentionally, no default value is defined.}

\item{sep}{\code{character} (optional) character used to separate input values
(default is nothing).}

\item{date}{\code{logical} Set TRUE to place the date before the message
(useful for logs or time consuming operations).}

\item{date_format}{\code{character} Format of the date (see \code{\link[base:strftime]{base::strftime()}})
for the definition of the format). The default format is
"\%Y-\%m-\%d \%H:\%M:\%S".}
}
\value{
message (in the defined format).
}
\description{
Print a message
}
\details{
Several functions print messages in different formats
(message, error, warning, cat, R output) and with different syntaxes
(concatenating parameters or accepting a single argument, appending
a new line, etc.).
This accessory function provides a common interface for different types:
several arguments are accepted and concatenated with the 'sep' argument;
the format is defined with the 'format' argument; a date is optionally
placed before the message.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
