% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projname.R
\name{projname}
\alias{projname}
\alias{projpar}
\title{Return the CRS name}
\usage{
projname(x, abort = FALSE)

projpar(x, par, abort = FALSE)
}
\arguments{
\item{x}{The CRS to be named (any \link{st_crs2} input is accepted).}

\item{abort}{logical: if TRUE, the function aborts in case an invalid
CRS is passed; if FALSE (default), the function returns NA,
and a warning is shown.}
}
\value{
A character with the CRS name, and an
attribute \code{crs} with the input projection checked using
\code{\link[sf:st_crs]{sf::st_crs()}}.
}
\description{
Return the name present in the WKT of the given CRS.
}
\details{
\code{\link[=projpar]{projpar()}} is deprecated.
}
\examples{
\donttest{
projname(4326)
}
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
