% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{s2-classes}
\alias{s2-classes}
\alias{as.s2dt}
\alias{as.s2list}
\title{Formats for SAFE archive lists}
\usage{
as.s2dt(x)

as.s2list(x)
}
\arguments{
\item{x}{Input object, which could be:
\itemize{
\item a named vector, in which elements are SciHub URLs and names are SAFE names;
\item a data frame or data.table, in which columns \code{name} and \code{url} must be present;
\item an object of class \code{s2list} or \code{s2dt}.
}}
}
\value{
\code{as.s2dt} returns an object of class \code{s2dt}: a data.table
in which column \code{name} contains SAFE names,
column \code{url} contains SciHub URLs
and additional columns include other product metadata.

\code{as.s2list} returns an object of class \code{s2list}: a named
vector in which elements are SciHub URLs and names are SAFE names,
eventually including some product metadata (stored as attributes).
}
\description{
Create or convert lists of SAFE archives from named vectors
or data frames.
}
\note{
License: GPL 3.0
}
\examples{
pos <- sf::st_sfc(sf::st_point(c(9.85,45.81)), crs = 4326)
time_window <- as.Date(c("2017-05-01", "2017-05-31"))

## Convert from an object s2list
list_safe <- s2_list(spatial_extent = pos, time_interval = time_window)
list_safe
class(list_safe) # default output class of s2_list()
as.character(list_safe) # convert to a simple named character
as.data.frame(list_safe) # convert to a data.frame (or to a data.table using as.data.table)
attr(list_safe, "date") # extract an hidden attribute from an s2list

## Convert from an object s2dt
out_dt <- as.s2dt(list_safe)
out_dt
class(out_dt)
as.character(out_dt) # convert to a simple named character
as.data.frame(out_dt) # convert to a data.frame (or to a data.table using as.data.table)
as.s2list(out_dt) # convert to an object s2list
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
