% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_calcindices.R
\name{s2_calcindices}
\alias{s2_calcindices}
\title{Compute maps of spectral indices}
\usage{
s2_calcindices(infiles, indices, outdir = ".", a = NA, b = NA, x = NA,
  format = NA, subdirs = NA, tmpdir = NA, compress = "DEFLATE",
  dataType = "Int16")
}
\arguments{
\item{infiles}{A vector of input filenames. Input files are paths
of BOA (or TOA) products already converted from SAFE format to a
format managed by GDAL (use \link{s2_translate} to do it);
their names must be in the fidolasen-S2 naming convention
(\link{s2_shortname}).}

\item{indices}{Character vector with the names of the required
indices. Values should be included in names corresponding to the
Abbreviations of the following indices:
\href{http://www.indexdatabase.de/db/is.php?sensor_id=96}{IDB}
FIXME the list of the accepted values is a subset; this reference
will be replaced with an internal html page integrated in the
shiny interface).}

\item{outdir}{(optional) Full name of the output directory where
the files should be created (default: current directory).
\code{outdir} can bot be an existing or non-existing directory (in the
second case, its parent directory must exists).
If it is a relative path, it is expanded from the common parent
directory of \code{infiles}.}

\item{format}{(optional) Format of the output file (in a
format recognised by GDAL). Default is the same format of input images
(or "GTiff" in case of VRT input images).}

\item{subdirs}{(optional) Logical: if TRUE, different indices are
placed in separated \code{outfile} subdirectories; if FALSE, they are placed in
\code{outfile} directory; if NA (default), subdirectories are created only if
more than a single spectral index is required.}

\item{tmpdir}{(optional) Path where intermediate VRT will be created.
Default is in a hidden subdirectory (called \code{.vrt}) of the common parent
directory of \code{infiles}. Set \code{tmpdir=tempdir()} if you do not want to
keep the intermediate files after reboot.}

\item{compress}{(optional) In the case a GTiff format is
present, the compression indicated with this parameter is used.}

\item{dataType}{(optional) Numeric datatype of the ouptut rasters:
if "Float32" or "Float64" is chosen, numeric values are not rescaled;
if "Int16" (default) or "UInt16", values are multiplicated by a 10000
scale factor.}
}
\value{
A vector with the names of the created products.
}
\description{
Create maps of a set of spectral indices. Since
\code{gdal_calc.py} is used to perform computations, output files
are physical rasters (no output VRT is allowed).
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
