% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_getMetadata.R
\name{s2_getMetadata}
\alias{s2_getMetadata}
\title{s2_getMetadata}
\usage{
s2_getMetadata(s2, info = "all")
}
\arguments{
\item{s2}{`character` or `osgeo.gdal.Dataset` This input parameter
can be the main path of a S2 file, the path of the xml with metadata,
th path of a single granule, the xml path of a single granule, or a
`osgeo.gdal.Dataset` object (obtained reading the product with python).}

\item{info}{`character` (optional) Vector with the list of the metadata
which should be provided. By default, all the metadata are provided.
By choosing "nameinfo", only the metadata obtained by scanning the file
name are provided (this is a bit faster). Alternatively, single
information can be asked.}
}
\value{
`list` list of the output metadata.
}
\description{
Get information from S2 file name or metadata
}
\details{
The function scan a Sentinel2 product (main path, granule path,
 main / granule xml file or GDAL object) to retrieve information about
 the product.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
