% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_gui.R
\name{s2_gui}
\alias{s2_gui}
\title{Launch the GUI for Sentinel-2 products}
\usage{
s2_gui(param_list = NULL, thunderforest_api = NA)
}
\arguments{
\item{param_list}{List of parameters for initialising the GUI values
(if empty, default values are used).}

\item{thunderforest_api}{Character value with the API for thunderforest
layers (now not used).}
}
\value{
A list of parameters.
}
\description{
Launch the GUI to set parameters for the processing
chain of Sentinel-2 products.
}
\note{
License: GPL 3.0
}
\references{
L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020).
"sen2r": An R toolbox for automatically downloading and preprocessing
Sentinel-2 satellite data. \emph{Computers & Geosciences}, 139, 104473. DOI:
\href{https://doi.org/10.1016/j.cageo.2020.104473}{10.1016/j.cageo.2020.104473},
URL: \url{http://sen2r.ranghetti.info/}.
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
