% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_list.R
\name{s2_list}
\alias{s2_list}
\title{Retrieve list of available products.}
\usage{
s2_list(spatial_extent = NULL, tile = NULL, orbit = NULL,
  time_interval = NULL, level = "auto", apihub = NULL, max_cloud = 110)
}
\arguments{
\item{spatial_extent}{A valid spatial object managed by \link[sprawl:get_extent]{sprawl::get_extent}}

\item{tile}{Single Sentinel-2 Tile string (5-length character)}

\item{orbit}{Single Sentinel-2 orbit number}

\item{time_interval}{a temporal vector (class \link{POSIXct} or
\link{Date}) of length 1 (specific day) or 2 (time interval).}

\item{level}{Character vector with one of the following:
- "auto" (default): check if level-2A is available on SciHub:
if so, list it; if not, list the corresponding level-1C
product
- "L1C": list available level-1C products
- "L2A": list available level-2A products}

\item{apihub}{Path of the "apihub.txt" file containing credentials
of scihub account. If NULL (default) the default credentials
(username "user", password "user") will be used.}

\item{max_cloud}{Integer number (0-100) containing the maximum cloud
level of the tiles to be listed (default: no filter).}
}
\value{
A vector of available products (being each element an URL,
and its name the product name).
}
\description{
The function retrieves the list of available Sentinel-2
products basing on search criteria. It makes use of
\href{https://github.com/ranghetti/s2download}{s2download}
python function only to retrieve the list of files, without
downloading and correcting them.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
pos <- sp::SpatialPoints(data.frame("x"=12.0,"y"=44.8), proj4string=sp::CRS("+init=epsg:4326"))
time_window <- as.Date(c("2017-05-01","2017-07-30"))
example_s2_list <- s2_list(spatial_extent=pos, tile="32TQQ", time_interval=time_window)
print(example_s2_list)
}
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
