% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_mask.R
\name{s2_mask}
\alias{s2_mask}
\title{Apply cloud masks}
\usage{
s2_mask(infiles, maskfiles, mask_type = "cloud_medium_proba",
  outdir = "./masked", format = NA, subdirs = NA, compress = "DEFLATE",
  parallel = FALSE, overwrite = FALSE)
}
\arguments{
\item{infiles}{A vector of input filenames. Input files are paths
of products already converted from SAFE format to a
format managed by GDAL (use \link{s2_translate} to do it);
their names must be in the fidolasen-S2 naming convention
(\link{s2_shortname}).}

\item{maskfiles}{A vector of filenames from which to take the
information about cloud coverage (for now, only SCL products
have been implemented). It is not necessary that \code{maskfiles}
elements strictly match \code{infiles} ones. Input files are paths
of products already converted from SAFE format to a
format managed by GDAL (use \link{s2_translate} to do it);
their names must be in the fidolasen-S2 naming convention
(\link{s2_shortname}).}

\item{mask_type}{Character vector which determines the type of
mask to be applied. Accepted values are:
\itemize{
\item "nodata": mask pixels checked as "No data" in the SCL product;
\item "cloud_high_proba": mask pixels checked as "No data" or
"Cloud (high probability)" in the SCL product;
\item "cloud_medium_proba": mask pixels checked as "No data" or
"Cloud (high or medium probability)" in the SCL product;
\item "cloud_low_proba": mask pixels checked as "No data" or
"Cloud (any probability)" in the SCL product;
\item "cloud_and_shadow": mask pixels checked as "No data",
"Cloud (any probability)" or "Cloud shadow" in the SCL product;
\item "cloud_shadow_cirrus": mask pixels checked as "No data",
"Cloud (any probability)", "Cloud shadow" or "Thin cirrus"
in the SCL product;
\item "opaque_clouds" (still to be implemented).
}}

\item{outdir}{(optional) Full name of the output directory where
the files should be created (default: "current directory"masked"
subdir of current directory).
\code{outdir} can bot be an existing or non-existing directory (in the
second case, its parent directory must exists).
If it is a relative path, it is expanded from the common parent
directory of \code{infiles}.}

\item{format}{(optional) Format of the output file (in a
format recognised by GDAL). Default is the same format of input images
(or "GTiff" in case of VRT input images).}

\item{subdirs}{(optional) Logical: if TRUE, different indices are
placed in separated \code{outfile} subdirectories; if FALSE, they are placed in
\code{outfile} directory; if NA (default), subdirectories are created only if
more than a single spectral index is required.}

\item{compress}{(optional) In the case a GTiff format is
present, the compression indicated with this parameter is used.}

\item{parallel}{(optional) Logical: if TRUE, masking is conducted using parallel
processing exploiting \link[raster:beginCluster]{raster::beginCluster}. This speeds-up the computation
for large rasters.
The number of cores is automatically determined; specifying it is also
possible (e.g. \code{parallel = 4}).
If FALSE (default), single core processing is used.}

\item{overwrite}{(optional) Logical value: should existing output files be
overwritten? (default: FALSE)}
}
\value{
A vector with the names of the created products.
}
\description{
Apply a cloud mask to a Sentinel-2 product. Since
\link{raster} functions are used to perform computations, output files
are physical rasters (no output VRT is allowed).
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
