% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_merge.R
\name{s2_merge}
\alias{s2_merge}
\title{Merge S2 tiles with the same date and orbit}
\usage{
s2_merge(infiles, outdir = ".", format = "VRT", compress = "DEFLATE",
  utmzone = "")
}
\arguments{
\item{infiles}{A vector of input filenames. Input files are paths
of products already converted from SAFE format to a format managed by
GDAL (use \link{s2_translate} to do it); their names must be in the
fidolasen-S2 naming convention (\link{s2_shortname}).}

\item{outdir}{(optional) Full name of the output directory where
the files should be created (default: current directory).
\code{outdir} can bot be an existing or non-existing directory (in the
second case, its parent directory must exists).}

\item{format}{(optional) Format of the output file (in a
format recognised by GDAL). Default value is "VRT" (Virtual Raster).}

\item{compress}{(optional) In the case a GTiff format is
chosen, the compression indicated with this parameter is used.}
}
\value{
A vector with the names of the merged products.
}
\description{
The function merge the input Sentinel-2 products with
the same date and orbit number. Outputs are a set of products in
the specified format.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
