% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_process_report.R
\name{s2_process_report}
\alias{s2_process_report}
\title{s2_process_report}
\usage{
s2_process_report(
  s2_list_ordered,
  s2names = NULL,
  pm = NULL,
  s2_list_cloudcovered = NA,
  s2_list_failed = NA,
  s2_list_cloud_ignored = NA,
  s2_list_failed_ignored = NA,
  download_only = FALSE
)
}
\arguments{
\item{s2_list_ordered}{list containing the lists of ordered/notordered lta
S2 images}

\item{s2names}{PARAM_DESCRIPTION}

\item{pm}{list containing sen2r processing parameters}

\item{s2_list_cloudcovered}{s2_list_cloudcovered}

\item{s2_list_failed}{s2_list_failed}

\item{s2_list_cloud_ignored}{s2_list_cloudcovered}

\item{s2_list_failed_ignored}{s2_list_failed}

\item{download_only}{logical, if TRUE, indicates that the processing to be
summarized only involved download (pm$proeprocess = FALSE)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function use to summarize results of sen2r execution at the
end of a processing.
}
\details{
DETAILS
}
