% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_sen2cor.R
\name{s2_sen2cor}
\alias{s2_sen2cor}
\title{Correct L1C products using sen2cor}
\usage{
s2_sen2cor(l1c_prodlist = NULL, l1c_dir = NULL, out_dir = NULL,
  n_procs = 1, overwrite = FALSE)
}
\arguments{
\item{l1c_prodlist}{'character' list of L1C product names to be corrected. They can be both
product names with full/relative path or only names of SAFE products (in this case, also
l1c_dir argument must be provided). SAFE products must be unzipped.
Note that, at this stage, all products must be in the same directory (this will be fixed).}

\item{l1c_dir}{'character' Full or relative path of input L1C products.}

\item{out_dir}{'character' Directory where output L2A products will be placed.}

\item{n_procs}{'integer' Number of processors to use (default is 1, single processor).}

\item{overwrite}{'logical' Should existing output L2A products be overwritten?
(default: FALSE)}
}
\description{
The function uses sen2cor from the docker to manually correct L1C products.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
pos <- sp::SpatialPoints(data.frame("x"=12.0,"y"=44.8), proj4string=sp::CRS("+init=epsg:4326"))
time_window <- as.Date(c("2017-05-01","2017-07-30"))
example_s2_list <- s2_list(spatial_extent=pos, tile="32TQQ", time_interval=time_window)
s2_download(example_s2_list, out_dir=tempdir())
s2_sen2cor(names(example_s2_list)[1], l1c_dir=tempdir(), out_dir=tempdir())
}
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
