% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_tiles.R
\name{s2_tiles}
\alias{s2_tiles}
\title{Load Sentinel-2 tiles}
\usage{
s2_tiles()
}
\value{
An sf spatial object containing the extent of the tiles.
}
\description{
Load the vector object of the Sentinel-2 tiles.
}
\note{
License: GPL 3.0
}
\examples{
# Retrieve all the tiles
s2tiles <- s2_tiles()

# Show a subset of all the tiles
s2tiles_it <- s2tiles[grep("^3[23]T", s2tiles$tile_id),]
s2_coords <- sf::st_coordinates(sf::st_centroid(s2tiles_it))
plot(s2tiles_it$geometry)
text(s2_coords[,1], s2_coords[,2], s2tiles_it$tile_id, cex=0.5)
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
