% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2buildvrt.R
\name{s2buildvrt}
\alias{s2buildvrt}
\title{s2buildvrt}
\usage{
s2buildvrt(infile, outdir = ".", utmzone = "")
}
\arguments{
\item{infile}{`character` Full path of the input SAFE folder (alternatively,
full path of the xml file of the product with metadata).}

\item{utmzone}{`character`(optional) UTM zone of output products (default:
the first one retrieved from input granules). Note that this function
does not perform reprojections: if no granules refer to the specified
UTM zone, no output is created.}

\item{outfile}{`character` (optional) Full path (or full name) where the
output vrt file should be created (default: current directory). If a
directory is provided, the file name will be the same of the SAFE input product.}
}
\description{
Build a vrt with all Sentinel2 bands
}
\details{
The function build a virtual raster from a Sentinel2 SAFE product.
 For now, only L1C and L2a with long name (< 2016/12/06) are recognised.
 Output vrt is at 10m resolution.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
