% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_is_online.R
\name{safe_is_online}
\alias{safe_is_online}
\title{Check if SAFE is available for download}
\usage{
safe_is_online(s2_prodlist = NULL, apihub = NA)
}
\arguments{
\item{s2_prodlist}{List of the products to be checked
(this must be the output of \link{s2_list} function).}

\item{apihub}{Path of the "apihub.txt" file containing credentials
of SciHub account.
If NA (default), the default location inside the package will be used.}
}
\value{
A logical vector of the same length and names of \code{s2_prodlist},
in which each element is TRUE if the corresponding SAFE archive is
available for download, FALSE if it is not or NA in case of errors with
the SAFE url.
}
\description{
The function checks if the required SAFE archives are
available for download, or if they have to be ordered from the Long Term
Archive.
}
\note{
License: GPL 3.0
}
\examples{
\donttest{
single_s2 <- paste0("https://scihub.copernicus.eu/apihub/odata/v1/",
  "Products(\\'c7142722-42bf-4f93-b8c5-59fd1792c430\\')/$value")
names(single_s2) <- "S2A_MSIL1C_20170613T101031_N0205_R022_T32TQQ_20170613T101608.SAFE"
# (this is equivalent to:
# single_s2 <- example_s2_list[1]
# where example_s2_list is the output of the example of the
# s2_list() function)

# Check for availability
safe_is_online(single_s2)
}
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
