% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safelist-class.R
\docType{class}
\name{safelist-class}
\alias{safelist-class}
\alias{safelist}
\title{Format for SAFE archive lists}
\description{
Create or convert lists of SAFE archives from named vectors
or data frames.
}
\note{
License: GPL 3.0
}
\examples{
pos <- sf::st_sfc(sf::st_point(c(9.85,45.81)), crs = 4326)
time_window <- as.Date(c("2017-05-01", "2017-05-31"))

## Create an object of class safelist
list_safe <- s2_list(spatial_extent = pos, time_interval = time_window)
list_safe
class(list_safe)
as.character(list_safe) # convert to a simple named character
as.data.frame(list_safe) # convert to a data.frame (or to a data.table using as.data.table)
attr(list_safe, "sensing_datetime") # extract an hidden attribute from a safelist
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
