% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scihub_login.R
\name{read_scihub_login}
\alias{read_scihub_login}
\alias{is_scihub_configured}
\alias{check_scihub_login}
\alias{check_scihub_connection}
\alias{write_scihub_login}
\title{Import / export / check SciHub username and password (deprecated)}
\usage{
read_scihub_login(apihub_path = NA)

is_scihub_configured()

check_scihub_login(username, password, service = "apihub")

check_scihub_connection(service = "apihub")

write_scihub_login(
  username,
  password,
  apihub_path = NA,
  check = TRUE,
  append = FALSE
)
}
\arguments{
\item{apihub_path}{deprecated}

\item{username}{deprecated}

\item{password}{deprecated}

\item{service}{deprecated}

\item{check}{deprecated}

\item{append}{deprecated}
}
\value{
deprecated

deprecated
}
\description{
These functions are deprecated and will be removed.
}
