% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_mask.R
\name{smooth_mask}
\alias{smooth_mask}
\title{Buffer clound masks}
\usage{
smooth_mask(inmask, binpaths, tmpdir, radius, buffer)
}
\arguments{
\item{inmask}{The path of the input 0-1 mask (where 0 represents the area
to be masked, 1 the cleen surface).}

\item{binpaths}{list of paths of binaries.}

\item{tmpdir}{(optional) Path where intermediate files (VRT) will be created.
Default is a temporary directory.}

\item{radius}{(optional) Numerical (positive): the size (in the unit of
\code{inmask}, typically metres) to be used as radius for the smoothing
(the higher it is, the more smooth the output mask will result).}

\item{buffer}{(optional) Numerical (positive or negative): the size of the
buffer (in the unit of \code{inmask}, typically metres) to be applied to the
masked area after smoothing it (positive to enlarge, negative to reduce).}
}
\value{
The path of the smoothed mask.
}
\description{
Internal function (used by \link{2s_mask}) which smooths
and buffers a 0-1 mask image in order to reduce the roughness of the mask
obtained from SCL classification (which is done pixel by pixel).
See details.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
