% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_text_2.R
\name{st_as_text_2}
\alias{st_as_text_2}
\title{Return WKT or WKT2 basing on the installed rgdal version}
\usage{
st_as_text_2(x)
}
\arguments{
\item{x}{object of class sfg, sfc or crs}
}
\value{
Well-known Text representation of simple feature geometry or
coordinate reference system
}
\description{
This is a convenience temporary function which returns the WKT
representation of a CRS, using \link[sf:st_as_text]{sf::st_as_text} in case PROJ < 3,
\link[rgdal:CRS]{rgdal::CRS} otherwise.
This has the advantage to perform precise transformations with PROJ >=3,
and to avoid conversion errors
(see https://rsbivand.github.io/ECS530_h19/ECS530_III.html).
This function will be deleted whenever \code{sf} will manage WKT2.
}
\examples{
sen2r:::st_as_text_2(st_crs(32632))
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
