% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_warnings.R
\name{suppress_warnings}
\alias{suppress_warnings}
\title{Selective suppress warnings}
\usage{
suppress_warnings(.expr, .f, ...)
}
\arguments{
\item{.expr}{Code to evaluate}

\item{.f}{A regular expression (which will be passed to \code{grepl()}).}
}
\value{
The warning message as character string, invisibly.
}
\description{
Suppress warnings matching particular regular expressions.
}
\details{
See https://stackoverflow.com/questions/16517795/selective-suppresswarnings-that-filters-by-regular-expression
}
