% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rgb_image.R
\name{add_rgb_image}
\alias{add_rgb_image}
\title{Add RGB product}
\usage{
add_rgb_image(s2_bands)
}
\arguments{
\item{s2_bands}{2-length list (one for TOA, one for BOA),
each element being a list of S2 bands, as defined in \link{s2_gui}.}
}
\description{
Modal dialog to define an RGB image.
}
\note{
License: GPL 3.0
}
\references{
L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020).
"sen2r": An R toolbox for automatically downloading and preprocessing
Sentinel-2 satellite data. \emph{Computers & Geosciences}, 139, 104473.
\doi{10.1016/j.cageo.2020.104473}, URL: \url{https://sen2r.ranghetti.info/}.
}
\author{
Luigi Ranghetti, phD (2019)
}
\keyword{internal}
