% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gdal.R
\name{check_gdal}
\alias{check_gdal}
\title{Check GDAL installation}
\usage{
check_gdal(abort = TRUE, force = FALSE)
}
\arguments{
\item{abort}{Logical parameter: if TRUE (default), the function aborts
in case no GDAL installation is found; if FALSE, a warning is shown
and FALSE is returned.}

\item{force}{(optional) Logical: if TRUE, install even if it is already
installed (default is FALSE).}
}
\value{
Logical: TRUE in case the installation is ok, FALSE if GDAL
is missing and abort=FALSE (otherwise, the function stops).
}
\description{
The function check that GDAL is installed and updated to
the minimum required version (2.1.3, since previous versions do not
manage SAFE format).
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{

# Remove previous GDAL information
options("gdalUtils_gdalPath" = NULL)
getOption("gdalUtils_gdalPath")

# Use function
check_gdal()
getOption("gdalUtils_gdalPath")
}
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
