% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sen2r_deps.R
\name{check_sen2r_deps}
\alias{check_sen2r_deps}
\title{Check package dependencies}
\usage{
check_sen2r_deps()
}
\value{
NULL (the function is called for its side effects)
}
\description{
The function allows to graphically check that all the
dependencies are installed.
}
\details{
This package needs some external dependencies to run:
\itemize{
\item Python
\item GDAL
\item Sen2Cor
}

This function opens a GUI which allows to check that these dependencies
are installed. This check is highly suggested before using the library for
the fist time, in order to avoid errors.
}
\note{
License: GPL 3.0
}
\examples{
if (interactive()) {
  check_sen2r_deps()
}
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
