% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_s2_paths.R, R/list_sen2r_paths.R
\name{compute_s2_paths}
\alias{compute_s2_paths}
\alias{list_sen2r_paths}
\title{Compute names of S2 file to be generated}
\usage{
compute_s2_paths(pm, s2_list_l1c, s2_list_l2a, paths, list_prods, out_ext,
  index_ext, tiles_ext, merged_ext, warped_ext, sr_masked_ext,
  force_tiles = FALSE, check_tmp = TRUE, ignorelist)

list_sen2r_paths(param_list = NULL, gui = NA, preprocess = TRUE,
  s2_levels = c("l1c", "l2a"), sel_sensor = c("s2a", "s2b"),
  online = TRUE, apihub = NA, downloader = "wget",
  overwrite_safe = FALSE, rm_safe = "no", step_atmcorr = "auto",
  timewindow = NA, timeperiod = "full", extent = NA, extent_name = "",
  s2tiles_selected = NA, s2orbits_selected = NA, list_prods = c("BOA"),
  list_indices = NA, index_source = "BOA", mask_type = NA,
  max_mask = 100, mask_smooth = 0, mask_buffer = 0,
  clip_on_extent = TRUE, extent_as_mask = FALSE, reference_path = NA,
  res = NA, res_s2 = "10m", unit = "Meter", proj = NA,
  resampling = "near", resampling_scl = "near", outformat = "GTiff",
  index_datatype = "Int16", compression = "DEFLATE", overwrite = FALSE,
  path_l1c = NA, path_l2a = NA, path_tiles = NA, path_merged = NA,
  path_out = NA, path_indices = NA, path_subdirs = TRUE,
  thumbnails = TRUE, parallel = TRUE, use_python = TRUE, tmpdir = NA,
  rmtmp = TRUE, log = NA)
}
\arguments{
\item{pm}{List of input parameters.}

\item{s2_list_l1c}{Names and paths of input SAFE level-1C products.}

\item{s2_list_l2a}{Names and paths of input SAFE level-2A products.}

\item{paths}{Named vector of required paths. It must contain elements
named \code{"out"}, \code{"indices"}, \code{"tiles"}, \code{"merged"} and \code{"warped"}.}

\item{list_prods}{Character vector with the values of the
products to be processed (accepted values: "TOA", "BOA", "SCL", "TCI").}

\item{out_ext}{Extension (character) of output products.}

\item{index_ext}{Extension (character) of index products.}

\item{tiles_ext}{Extension (character) of tiled products.}

\item{merged_ext}{Extension (character) of merged products.}

\item{warped_ext}{Extension (character) of warped products.}

\item{sr_masked_ext}{Extension (character) of masked products of SR products.}

\item{force_tiles}{(optional) Logical: passed to \link{safe_shortname} (default: FALSE).}

\item{check_tmp}{(optional) Logical: if TRUE (default), temporary files
are also searched when \code{_exi} names are computed;
if FALSE, only non temporary files are searched.}

\item{ignorelist}{Vector of output files to be ignored.}

\item{param_list}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{gui}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{preprocess}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{s2_levels}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{sel_sensor}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{online}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{apihub}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{downloader}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{overwrite_safe}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{rm_safe}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{step_atmcorr}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{timewindow}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{timeperiod}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{extent}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{extent_name}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{s2tiles_selected}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{s2orbits_selected}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{list_indices}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{index_source}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{mask_type}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{max_mask}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{mask_smooth}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{mask_buffer}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{clip_on_extent}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{extent_as_mask}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{reference_path}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{res}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{res_s2}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{unit}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{proj}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{resampling}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{resampling_scl}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{outformat}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{index_datatype}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{compression}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{overwrite}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{path_l1c}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{path_l2a}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{path_tiles}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{path_merged}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{path_out}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{path_indices}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{path_subdirs}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{thumbnails}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{parallel}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{use_python}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{tmpdir}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{rmtmp}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{log}{\code{sen2r} argument (refer to \link{sen2r} documentation).}

\item{list_prods}{\code{sen2r} argument (refer to \link{sen2r} documentation).}
}
\value{
A list, containing the following vectors:
\code{tiles_names_exp}, \code{merged_names_exp}, \code{warped_names_exp},
\code{masked_names_exp}, \code{out_names_exp}, \code{indices_names_exp},
\code{indices_names_new}, \code{out_names_req}, \code{out_names_new},
\code{masked_names_new}, \code{warped_names_req}, \code{warped_names_new},
\code{merged_names_req}, \code{merged_names_new}, \code{tiles_names_req}, \code{tiles_names_new},
\code{safe_names_l1c_req}, \code{safe_names_l2a_req}.
}
\description{
\code{compute_s2_paths} is an internal function
(to be used within \code{\link[=sen2r]{sen2r()}})
which computes the names of the required output image files
(see details).
The function was splitted from \code{\link[=sen2r]{sen2r()}} because this code
is called twice (and to shorten the main function).

\code{list_sen2r_paths} is a wrapper of \link{sen2r}, which runs
\link{sen2r} until \link{compute_s2_paths} is called, returning the same list.
It is a simple way to call \link{compute_s2_paths} with the same arguments
if \link{sen2r}.
}
\details{
\code{compute_s2_paths} is structured in the following way:
\enumerate{
\item Retrieve the file names expected to be present at the
end of the processing chain (suffix \code{_exp}):
\itemize{
\item \code{tiles_names_}
\item \code{merged_names_}
\item \code{warped_names_}
\item \code{masked_names_}
\item \code{out_names_} (= \code{warped_names_} or \code{subset(warped_names_)+masked_names_})
\item \code{indices_names_}
}
\item Compute the file names expected to be created
(suffixes \code{_req} and \code{_new}, see below)
(this operation is done in reverse order).

Meaning of the suffixes \code{_exp}, \code{_req} and \code{_new}
(here and for all the script):
\itemize{
\item \code{_exp}: full names of the files expected to be present at the
end of the processing chain (already existing or not);
\item \code{_req}: names of the files required for the next step
(e.g. tiles_names_req are required to perform s2_merge())
\item \code{_new}: names of the required files not existing yet (expected
to be created).
}
}

With \code{overwrite=TRUE}, all these vectors are equal
(e.g. \code{merged_names_exp = merged_names_req = merged_names_new}),
because all is overwritten.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
