% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indices_db.R
\name{create_indices_db}
\alias{create_indices_db}
\title{Create the indices database}
\usage{
create_indices_db(xslt_path = NA, json_path = NA)
}
\arguments{
\item{xslt_path}{(optional) The path where to install \code{xsltml},
an external \code{xsltproc} script used to convert MathML index formulas
to LaTeX (default: a subdirectory of the package).}

\item{json_path}{(optional) The path of the output JSON file.
\emph{Warning}: to create a file which wil be usable by the package,
this option must be left to NA (default location is within the
package installation). Edit this only to create the file in another
place for external use.}
}
\description{
The internal function checks if idices.json (the
database of spectral indices) already exists; if not, it
downloads source files and creates it.
Since this function depends on xsltproc executable (available
inly for Linux), this function can be used only from from
Linux. It is not necessary, since a indices.json file is
present in the package.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
