% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_equal_crs.R
\name{expect_equal_crs}
\alias{expect_equal_crs}
\title{Compare two non-null CRS}
\usage{
expect_equal_crs(crs1, crs2)
}
\arguments{
\item{crs1}{CRS 1 to compare}

\item{crs2}{CRS 2 to compare}
}
\value{
\code{testthat::expect_equal()} output.
}
\description{
Internal function: convenience function to compare two non-null
CRS in tests without using EPSG (so usable with rgdal >= 1.5).
}
\note{
License: GPL 3.0
}
\references{
L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020).
"sen2r": An R toolbox for automatically downloading and preprocessing
Sentinel-2 satellite data. \emph{Computers & Geosciences}, 139, 104473.
\doi{10.1016/j.cageo.2020.104473}, URL: \url{https://sen2r.ranghetti.info/}.
}
\author{
Luigi Ranghetti, phD (2020)
}
\keyword{internal}
