% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidolasen_s2.R
\name{fidolasen_s2}
\alias{fidolasen_s2}
\title{Find, download and preprocess Sentinel-2 images}
\usage{
fidolasen_s2(param_list = NULL, preprocess = NA, s2_levels = NA,
  sel_sensor = NA, online = NA, overwrite_safe = NA, rm_safe = NA,
  step_atmcorr = NA, timewindow = NA, timeperiod = NA, extent = NA,
  s2tiles_selected = NA, s2orbits_selected = NA, extent_as_mask = NA,
  list_prods = NA, list_indices = NA, index_source = NA, mask_type = NA,
  reference_path = NA, res = NA, res_s2 = NA, unit = NA, proj = NA,
  resampling = NA, resampling_scl = NA, outformat = NA,
  compression = NA, overwrite = NA, path_l1c = NA, path_l2a = NA,
  path_tiles = NA, path_merged = NA, path_out = NA, path_indices = NA,
  path_subdirs = NA)
}
\arguments{
\item{param_list}{List of input parameters: it can be both an
R list or the path of a JSON file.
If some parameters are passed both as elements of \code{param_list}
and as function arguments, the values passed as function
arguments are considered.
If some parameters are missing in \code{param_list} and are not
provided as arguments, default values will be used.
Use the function \code{\link[=s2_gui]{s2_gui()}} to create a complete list of
parameters.}

\item{preprocess}{=NA,}

\item{s2_levels}{=NA,}

\item{sel_sensor}{=NA,}

\item{online}{=NA,}

\item{overwrite_safe}{=NA,}

\item{rm_safe}{=NA,}

\item{step_atmcorr}{=NA,}

\item{timewindow}{=NA,}

\item{timeperiod}{=NA,}

\item{extent}{=NA,}

\item{s2tiles_selected}{=NA,}

\item{s2orbits_selected}{=NA,}

\item{extent_as_mask}{=NA,}

\item{list_prods}{=NA,}

\item{list_indices}{=NA,}

\item{index_source}{=NA,}

\item{mask_type}{=NA,}

\item{reference_path}{=NA,}

\item{res}{=NA,}

\item{res_s2}{=NA,}

\item{unit}{=NA,}

\item{proj}{=NA,}

\item{resampling}{=NA,}

\item{resampling_scl}{=NA,}

\item{outformat}{=NA,}

\item{compression}{=NA,}

\item{overwrite}{=NA,}

\item{path_l1c}{=NA,}

\item{path_l2a}{=NA,}

\item{path_tiles}{=NA,}

\item{path_merged}{=NA,}

\item{path_out}{=NA,}

\item{path_indices}{=NA,}

\item{path_subdirs}{=NA,}
}
\description{
The function is a wrapper to perform the entire
processing chain to find, download and pre-process Sentinel-2
data. Input is a set of parameters that can be passed with a
list or file (parameter \code{param_list}) or singularly (see the
descriptions of all the other parameters).
}
