% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fidolasen_s2.R
\name{fidolasen_s2}
\alias{fidolasen_s2}
\title{Find, download and preprocess Sentinel-2 images}
\usage{
fidolasen_s2(param_list = NULL, gui = NA, preprocess = NA,
  s2_levels = NA, sel_sensor = NA, online = NA, overwrite_safe = NA,
  rm_safe = NA, step_atmcorr = NA, timewindow = NA, timeperiod = NA,
  extent = NA, s2tiles_selected = NA, s2orbits_selected = NA,
  list_prods = NA, list_indices = NA, index_source = NA, mask_type = NA,
  clip_on_extent = NA, extent_as_mask = NA, reference_path = NA,
  res = NA, res_s2 = NA, unit = NA, proj = NA, resampling = NA,
  resampling_scl = NA, outformat = NA, compression = NA, overwrite = NA,
  path_l1c = NA, path_l2a = NA, path_tiles = NA, path_merged = NA,
  path_out = NA, path_indices = NA, path_subdirs = NA)
}
\arguments{
\item{param_list}{(optional) List of input parameters:
it can be both an R list or the path of a JSON file.
If some parameters are passed both as elements of \code{param_list}
and as function arguments, the values passed as function
arguments are considered.
If some parameters are missing in \code{param_list} and are not
provided as arguments, default values will be used.
Use the function \code{\link[=s2_gui]{s2_gui()}} to create a complete list of
parameters.
If \code{param_list} is NULL (default), values given with the
parameters below (or default values for parameters not
provided) are used.}

\item{gui}{(optional) Logical: if TRUE, function \code{\link[=s2_gui]{s2_gui()}} is
launched before starting to process in order to set or load parameters;
if FALSE, the function uses parameters passed with \code{param_list} or
with other function arguments. Default is FALSE if \code{param_list} is not
NULL, TRUE elsewhere.}

\item{preprocess}{(optional) Logical: TRUE (default) to perform also
preprocessing steps, FALSE not to (do only find, download
and atmospheric correction).}

\item{s2_levels}{(optional) Character vector of length 1 or 2, with
Sentinel-2 levels required for processing steps or as output.
Accepted values: "l1c" and "l2a"; default: "l2a".}

\item{sel_sensor}{(optional) Character vector of length 1 or 2, with
Sentinel-2 sensors to be used.
Accepted values: "s2a" and "s2b"; default: c("s2a","s2b").}

\item{online}{(optional) Logical: TRUE (default) to search for available
products on SciHub (and download if needed); FALSE to work
only with already downloaded SAFE products.}

\item{overwrite_safe}{(optional) Logical: TRUE to overwrite existing
products with products found online or manually corrected,
FALSE (default) to skip download and atmospheric correction for
products already existing.}

\item{rm_safe}{(optional) Character: should SAFE products be deleted after
preprocessing? "yes" means to delete all SAFE; "no" (default)
not to delete; "l1c" to delete only Level-1C products.}

\item{step_atmcorr}{(optional) Character vector to determine how to obtain
Level-2A SAFE products:
\itemize{
\item "auto" (default) means that L2A is first
searched on SciHub: if found, it is dowloaded, if not, the
corresponding Level-1C is downloaded and sen2cor is used to
produce L2A;
\item "scihub" means that sen2cor is always used from L1C products
downloaded from SciHub;
\item "l2a" means that they are downloaded if available on SciHub,
otherwise they are skipped (sen2cor is never used);
\item "no" means that L1C are not considered (processing chain
makes use only of L1C products).
}}

\item{timewindow}{(optional) Temporal window for querying: Date object
of length 1 (single day) or 2 (time window). Default is NA in online mode
(meaning that no filters are used, and all found images are processed); in
online mode, default is to process last 90 days.
Only in offline mode, it is possible to pass NA not to
filter by sensing time.}

\item{timeperiod}{(optional) Character:
\itemize{
\item "full" (default) means that all
the images included in the time window are considered;
\item "seasonal" means that only the single seasonal periods in the
window are used (i.e., with a time window from 2015-06-01 to
2017-08-31, the periods 2015-06-01 to 2015-08-31, 2016-06-01
to 2016-08-31 and 2017-06-01 to 2017-08-31 are considered).
}}

\item{extent}{(optional) Spatial extent on which to clip products, in
geojson format. Default is NA for offline mode (meaning no extent:
all found tiles are entirely used); in online mode, a sample extent is used.}

\item{s2tiles_selected}{(optional) Character vector with the Sentinel-2
tiles to be considered (default is NA, meaning all the tiles).}

\item{s2orbits_selected}{(optional) Character vector with the Sentinel-2
orbits to be considered (still to be implemented; for now,
all the accepted values are listed).}

\item{list_prods}{(optional) Character vector with the values of the
products to be processed (accepted values: "TOA", "BOA", "SCL",
"TCI"). Default is "BOA".}

\item{list_indices}{(optional) Character vector with the values of the
spectral indices to be computed. Default is no one (NA).}

\item{index_source}{(optional) Character value: if "BOA" (default), indices
are computed from BOA values; if "TOA", non corrected reflectances
are instead used (be careful to use this setting!).}

\item{mask_type}{(optional) Character value which determines the categories
in the Srface Classification Map to be masked (see \code{\link[=s2_mask]{s2_mask()}}
for the accepted values). Default (NA) is not to mask.}

\item{clip_on_extent}{(optional) Logical: if TRUE (default), output products
and indices are clipped to the selected extent (and resampled/reprojected);
if FALSE, the geometry and extension of the tiles is maintained.}

\item{extent_as_mask}{(optional) Logical: if TRUE, pixel values outside
the \code{extent} polygon are set to NA; if FALSE (default), all the values
within the bounding box are maintained.}

\item{reference_path}{(optional) Path of the raster file to be used as a
reference grid. If NA (default), no reference is used.}

\item{res}{(optional) Numerifc vector of length 2 with the x-y resolution
for output products. Default: c(10,10). NA means that the resolution
is keeped as native.}

\item{res_s2}{(optional) Character value corresponding to the native Sentinel-2
resolution to be used. Accepted values are "10m" (default), "20m"
and "60m".}

\item{unit}{(optional) Character value corresponding to the unit of measure
with which to interpret the resolution (for now, only "Meter" -
the default value - is supported).}

\item{proj}{(optional) Character string with the pro4string of the output
resolution. default value (NA) means not to reproject.}

\item{resampling}{(optional) Resampling method (one of the values supported
by \code{gdal_translate}: "near" (default), "bilinear", "cubic",
"cubicspline", "lanczos", "average" or "mode").}

\item{resampling_scl}{(optional) Resampling method for categorical products
(for now, only SCL): one among "near" (default) and "mode".}

\item{outformat}{(optional) Format of the output file (in a
format recognised by GDAL). Default is "GTiff".}

\item{compression}{(optional) In the case GTiff is chosen as
output format, the compression indicated with this parameter is
used (default is "DEFLATE").}

\item{overwrite}{(optional) Logical value: should existing output
files be overwritten? (default: FALSE).}

\item{path_l1c}{(optional) Path of the directory in which Level-1C SAFE
products are searched and/or downloaded. If not provided (default), a
temporary directory is used.}

\item{path_l2a}{(optional) Path of the directory in which Level-2A SAFE
products are searched, downloaded and/or generated. If not provided
(default), a temporary directory is used.}

\item{path_tiles}{(optional) Path of the directory in which Sentinel-2
tiles (as generated by \link{s2_translate}) are searched and/or generated.
If not provided (default), a temporary directory is used, and files
are generated as virtual rasters; otherwise, they are generated in
the format specified with \code{outformat} parameter.}

\item{path_merged}{(optional) Path of the directory in which Sentinel-2
tiles merged by orbit (as generated by \link{s2_merge}) are searched and/or
generated.
If not provided (default), a temporary directory is used, and files
are generated as virtual rasters; otherwise, they are generated in
the format specified with \code{outformat} parameter.}

\item{path_out}{(optional) Path of the directory in which Sentinel-2
output products are searched and/or generated.
If not provided (default), a temporary directory is used.}

\item{path_indices}{(optional) Path of the directory in which files of
spectral indices are searched and/or generated.
If not provided (default), \code{path_out} is used.}

\item{path_subdirs}{(optional) Logical: if TRUE (default), a directory
for each output product or spectral index is generated within
\code{path_tiles}, \code{path_merged}, \code{path_out} and \code{path_indices}; if FALSE,
products are put directly within them.}
}
\description{
The function is a wrapper to perform the entire
processing chain to find, download and pre-process Sentinel-2
data. Input is a set of parameters that can be passed with a
list or file (parameter \code{param_list}) or singularly (see the
descriptions of all the other parameters).
}
