% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalUtil.R
\name{gdalUtil}
\alias{gdalUtil}
\title{Interface to GDAL Python-based utilities}
\usage{
gdalUtil(
  util,
  source,
  destination,
  options = character(0),
  quiet = FALSE,
  formula = character(0),
  processing = character(0),
  colorfilename = character(0)
)
}
\arguments{
\item{util}{Character: one among \code{"info"}, \code{"translate"}, \code{"warp"},
\code{"demprocessing"} ,\code{"buildvrt"} (C-based),
\code{"calc"} and \code{"fillnodata"} (Python-based).
Other utilities were not implemented, since they are not used by sen2r.}

\item{source}{path of input layer(s); for \code{calc} this can be more than one.}

\item{destination}{Path of the output layer.}

\item{options}{Character vector with GDAL options.}

\item{quiet}{Logical: if TRUE, suppress printing of output for info
(this argument is ignored in case package \code{sys} is not installed).}

\item{formula}{(for \code{util = "calc"}) Calculation in gdalnumeric syntax using
\code{+}, \code{-}, \code{/}, \code{*}, or any \code{numpy} array functions (i.e. \code{log10()}).}

\item{processing}{Character: processing options for \code{util = "demprocessing"}.}

\item{colorfilename}{Character: name of color file for \code{util = "demprocessing"}
(mandatory if \code{processing="color-relief"}).}
}
\value{
A logical (invisible) indicating success (i.e., TRUE);
in case of failure, an error is raised.
}
\description{
This accessory function interfaces with GDAL
utilities (sen2r must be interfaced with a runtime GDAL
environment, see \code{check_gdal()}).
Python-based utilities are always called from an existing runtime GDAL
environment (see \code{check_gdal()});
C-based ones are called using \code{sf::gdal_utils()}.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
