% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gipp.R
\name{read_gipp}
\alias{read_gipp}
\alias{set_gipp}
\title{Manage GIPP parameters for Sen2Cor}
\usage{
read_gipp(gipp_names, gipp_path_in = NA)

set_gipp(gipp, gipp_path_in = NA, gipp_path_out = gipp_path_in)
}
\arguments{
\item{gipp_names}{Character vector with the names of the parameters
to be read.}

\item{gipp_path_in}{Character path of the GIPP XML file to be read.
If NA (default), the default sen2r GIPP file is used.}

\item{gipp}{List of Ground Image Processing Parameters (GIPP),
in the form \code{parameter_name = "value"}, where
\code{parameter_name} is the name of the parameter as specified in the
\code{L2A_GIPP.xml} file of the used Sen2Cor version (case insensitive), and
\code{"value"} is the character value which the user wants to set
(notice that, in the case the user wants to specify the value \code{NONE},
both \code{"NONE"} and \code{NA} can be used, but not \code{NULL}, which has the effect
to maintain the value specified in the XML file).
Elements whose name is missing in the XML file are skipped.}

\item{gipp_path_out}{Character path of the output GIPP XML file.
By default it is equal to \code{gipp_path_in} (edited values are overwritten).}
}
\value{
\code{read_gipp()} returns a named list of GIPP with the required parameters
(values not found in the XML are skipped).

\code{set_gipp()} returns NULL (the function is called for its side effects).
}
\description{
\code{read_gipp()} reads Ground Image Processing Parameters (GIPP)
from an XML file.

\code{set_gipp()} modify values of a list of Ground Image Processing
Parameters (GIPP) in an XML file.
}
\note{
License: GPL 3.0
}
\examples{
\donttest{
# Read default values
read_gipp(c("dem_directory", "dem_reference"))
# Edit one value and save the output as temporary file
set_gipp(list(DEM_Directory = NA), gipp_path_out = gipp_temp <- tempfile())
# Read the parameters in the created temporary files
read_gipp(c("DEM_Directory", "DEM_Reference"), gipp_path_in = gipp_temp)
# Read the original (not edited) XML file
read_gipp(c("DEM_Directory", "DEM_Reference"))
}
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
