% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gipp.R
\name{read_gipp}
\alias{read_gipp}
\alias{set_gipp}
\alias{reset_gipp}
\title{Manage GIPP parameters for Sen2Cor}
\usage{
read_gipp(gipp_names, gipp_path = NA)

set_gipp(gipp = list(), gipp_path = NA, use_dem = NA)

reset_gipp(gipp_path = NA)
}
\arguments{
\item{gipp_names}{Character vector with the names of the parameters
to be read.}

\item{gipp_path}{Character path of the output GIPP XML file.
If NA (default), the default sen2r GIPP path is used.}

\item{gipp}{Ground Image Processing Parameters (GIPP).
It is possible to specify both the path of an existing XML file
or a list of parameters in the form \code{parameter_name = "value"}, where
\code{parameter_name} is the name of the parameter as specified in the
\code{L2A_GIPP.xml} file of the used Sen2Cor version (case insensitive), and
\code{"value"} is the character value which the user wants to set
(notice that, in the case the user wants to specify the value \code{NONE},
both \code{"NONE"} and \code{NA} can be used, but not \code{NULL}, which has the effect
to maintain the value specified in the XML file).
Elements whose name is missing in the XML file are skipped.}

\item{use_dem}{Logical: if TRUE, a DEM is set for being used for topographic
corection in the XML specified with argument \code{gipp_path};
if FALSE, no DEM is set for being used;
if NA (default), the existing option is maintained.
Notice that, in case \code{use_dem = TRUE} and a DEM directory is already set
in the XML file, no changes are performed;
conversely; if no directories were specified, a subdirectory \code{"srtm90"}
of the default sen2r directory is set.
To set another directory, use argument \code{gipp} in the form
\code{gipp = list(DEM_Directory = tempdir(), ...)} (replacing \code{tempdir()} with
the desired path).}
}
\value{
\code{read_gipp()} returns a named list of GIPP with the required parameters
(values not found in the XML are skipped).

\code{set_gipp()} and \code{reset_gipp()} return NULL
(functions are called for their side effects).
}
\description{
\code{read_gipp()} reads Ground Image Processing Parameters (GIPP)
from an XML file.

\code{set_gipp()} modifies values of a list of GIPP in an XML file.

\code{reset_gipp()} restores default GIPP values (using values
set in the \code{L2A_GIPP.xml} default Sen2Cor file).
}
\note{
License: GPL 3.0
}
\examples{
\donttest{
# Read default values
read_gipp(c("dem_directory", "dem_reference"))
# Edit one value and save the output as temporary file
set_gipp(list(DEM_Directory = NA), gipp_path = gipp_temp <- tempfile())
# Read the parameters in the created temporary files
read_gipp(c("DEM_Directory", "DEM_Reference"), gipp_path = gipp_temp)
# Read the original (not edited) XML file
read_gipp(c("DEM_Directory", "DEM_Reference"))
# Reset to default Sen2Cor GIPP values
reset_gipp(gipp_path = gipp_temp)
# Read again values
read_gipp(c("DEM_Directory", "DEM_Reference"), gipp_path = gipp_temp)
}
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
