% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_python.R
\name{init_python}
\alias{init_python}
\title{Initialise python}
\usage{
init_python()
}
\value{
A list with the required modules.
}
\description{
Internal function to check that the python dependencies
required in the package are present; if so, load the base modules.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
