% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_python.R
\name{init_python}
\alias{init_python}
\title{Initialise python}
\usage{
init_python()
}
\value{
NULL (the function is called for its side effects)
}
\description{
Internal function to set the environmental variables
required to run Python-based GDAL utilities on Windows.
}
\note{
License: GPL 3.0
}
\references{
L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020).
"sen2r": An R toolbox for automatically downloading and preprocessing
Sentinel-2 satellite data. \emph{Computers & Geosciences}, 139, 104473.
\doi{10.1016/j.cageo.2020.104473}, URL: \url{http://sen2r.ranghetti.info/}.
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
