% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_aria2.R
\name{install_aria2}
\alias{install_aria2}
\title{Download and install aria2.}
\usage{
install_aria2(aria2_dir, force = FALSE)
}
\arguments{
\item{aria2_dir}{Path where aria2 executable will be installed.}

\item{force}{(optional) Logical: if TRUE, install even if it is already
installed (default is FALSE).}
}
\value{
The path of aria2
}
\description{
This function download and install standalone version of
\href{https://aria2.github.io}{aria2} for Windows.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
# Run only on Windows
install_aria2(aria2_dir = tempdir())
# ( use a non-temporary folder path instead of tempdir() )
}
}
\references{
L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020).
"sen2r": An R toolbox for automatically downloading and preprocessing
Sentinel-2 satellite data. \emph{Computers & Geosciences}, 139, 104473.
\doi{10.1016/j.cageo.2020.104473}, URL: \url{https://sen2r.ranghetti.info/}.
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
