% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_s2download.R
\name{install_s2download}
\alias{install_s2download}
\title{install_s2download}
\usage{
install_s2download(inst_path = NA)
}
\arguments{
\item{inst_path}{\code{character} Path where s2download will be cloned
(default: a subdirectory of the RSPrePro package).}
}
\description{
Clone s2download and install sen2cor docker.
}
\details{
s2download is a collection of python scripts used to download
and correct Sentinel-2 images, and it is required by RSPrePro.
This function clones them and installs a docker with sen2cor.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
