% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sen2cor.R
\name{install_sen2cor}
\alias{install_sen2cor}
\alias{link_sen2cor}
\title{Download and install (or link) Sen2Cor}
\usage{
install_sen2cor(
  sen2cor_dir = NA,
  version = "2.5.5",
  gipp = NULL,
  use_dem = NA,
  force = FALSE
)

link_sen2cor(sen2cor_dir, gipp = NULL, use_dem = NA)
}
\arguments{
\item{sen2cor_dir}{Path where sen2cor will be installed or searched
(by default it is a subdirectory \code{"sen2cor"} of the default sen2r directory).}

\item{version}{(optional) Character: Sen2Cor version (one among
'2.5.5' - default - and '2.8.0').}

\item{gipp}{(optional) Ground Image Processing Parameters (GIPP)
to be passed to Sen2Cor (see \code{\link[=set_gipp]{set_gipp()}} for further details).
Default behaviour is to maintain default values (with the exception of
parameter \code{DEM_Directory}, which is managed by argument \code{use_dem}).}

\item{use_dem}{(optional) Logical, determining if Sen2Cor has to be set
to use a Digital Elevation Model for topographic correction (see
\code{\link[=sen2cor]{sen2cor()}} for further details).}

\item{force}{(optional) Logical: if TRUE, installs Sen2Cor even if it is
already found in \code{sen2cor_dir} (default is FALSE).}
}
\value{
NULL (the function is called for its side effects)
}
\description{
\code{\link[=install_sen2cor]{install_sen2cor()}} downloads and installs a standalone version of
\href{http://step.esa.int/main/third-party-plugins-2/sen2cor}{Sen2Cor}.

\code{link_sen2cor()} links an existing standalone version of
\href{http://step.esa.int/main/third-party-plugins-2/sen2cor}{Sen2Cor} to sen2r.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
install_sen2cor(sen2cor_dir = tempdir(), use_dem = TRUE)
# ( use a non-temporary folder path instead of tempdir() )
}
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
