% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sen2cor.R
\name{install_sen2cor}
\alias{install_sen2cor}
\title{Download and install sen2cor.}
\usage{
install_sen2cor(sen2cor_dir = NA, force = FALSE)
}
\arguments{
\item{sen2cor_dir}{(optional) Path where sen2cor will be installed
(default: a subdirectory of the package path).}

\item{force}{(optional) Logical: if TRUE, install even if it is already
installed (default is FALSE).}
}
\description{
This function download and install standalone version of
\href{http://step.esa.int/main/third-party-plugins-2/sen2cor}{sen2cor 2.5.5}.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
install_sen2cor()
}
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
