% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_extent.R
\name{load_extent_bbox}
\alias{load_extent_bbox}
\alias{load_extent_vectfile}
\alias{load_extent_draw}
\title{Insert an extent}
\usage{
load_extent_bbox()

load_extent_vectfile()

load_extent_draw(extent_ns_name)
}
\arguments{
\item{extent_ns_name}{Name of the namespace to be used}
}
\description{
Internal functions and modal dialogs to specify an extent
in the GUI.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
