% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nn}
\alias{nn}
\title{Replace NULL with character()}
\usage{
nn(x)
}
\arguments{
\item{x}{Input variable}
}
\value{
\code{character(0)} if \code{x==NULL}, \code{x} elsewhere
}
\description{
Internal function: return character(0) instead of NULL.
This is sometimes needed not to return error when applying some functions.
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
basename(NULL) # error
basename(character()) # ok
basename(nn(NULL)) # ok
}
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
