% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_path.R
\name{normalize_path}
\alias{normalize_path}
\title{Express file paths in canonical Form depending on the operating system}
\usage{
normalize_path(path, ...)
}
\arguments{
\item{path}{character vector of file paths}

\item{...}{additional parameters passed to \link{normalizePath} (i.e. \code{mustWork}).}
}
\value{
The paths normalized.
}
\description{
Accessory function wrapper for \code{normalizePath()} in Linux
and \code{shortPathName(normalizePath())} in Windows.
}
\note{
License: GPL 3.0
}
\references{
L. Ranghetti, M. Boschetti, F. Nutini, L. Busetto (2020).
"sen2r": An R toolbox for automatically downloading and preprocessing
Sentinel-2 satellite data. \emph{Computers & Geosciences}, 139, 104473.
\doi{10.1016/j.cageo.2020.104473}, URL: \url{http://sen2r.ranghetti.info/}.
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
