% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_check.R
\name{path_check}
\alias{path_check}
\title{Check a path}
\usage{
path_check(path, mustbe_empty = FALSE, mustbe_writable = TRUE)
}
\arguments{
\item{path}{\code{string} fullpath to a folder}

\item{mustbe_empty}{\code{logical} if TRUE, accept only empty directories}

\item{mustbe_writable}{\code{logical} if TRUE, accept only writable directories}
}
\description{
Accessory functions to check that a directory exists and
is writable (to be used in a Shiny context)
}
\author{
Luigi Ranghetti, PhD (2019) \email{ranghetti.l@irea.cnr.it}

Lorenzo Busetto, PhD (2019) \email{lbusett@gmail.com>}
}
