% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projname.R
\name{projname}
\alias{projname}
\title{Return the name of a projection}
\usage{
projname(proj4string, abort = FALSE)
}
\arguments{
\item{proj4string}{The proj4string to be named (a character or a \link{CRS}
object).}

\item{abort}{logical: if TRUE, the function aborts in case an invalid
proj4string is passed; if FALSE (default), the function returns NA,
and a warning is shown.}
}
\value{
A character with the name of the projection (and an attribute
\code{proj4string} with the input projection checked using
\link[sprawl:check_proj4string]{sprawl::check_proj4string}).
}
\description{
Return the projection name of a recognised proj4string.
}
\note{
Python is needed.
}
\examples{
projname("+init=epsg:4326")
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
