% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projpar.R
\name{projpar}
\alias{projpar}
\alias{projname}
\title{Return a parameter used in a WKT projection}
\usage{
projpar(x, par, abort = FALSE)

projname(x, abort = FALSE)
}
\arguments{
\item{x}{The CRS to be named (any \link{st_crs2} input is accepted).}

\item{par}{Character corresponding to the parameter name.}

\item{abort}{logical: if TRUE, the function aborts in case an invalid
CRS is passed; if FALSE (default), the function returns NA,
and a warning is shown.}
}
\value{
A character with the content of the parameter (NULL if the
parameter is not recognised) or the name of the projection, and an
attribute \code{crs} with the input projection checked using
\code{\link[sf:st_crs]{sf::st_crs()}}.
}
\description{
Return the value of a parameter (or the name) present in
the WKT of the given CRS
}
\note{
Python is needed.
}
\examples{
\donttest{
projpar(4326, "Unit")
}
\donttest{
projname(4326)
}
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
