% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_metadata.R
\name{raster_metadata}
\alias{raster_metadata}
\title{Get metadata from raster paths}
\usage{
raster_metadata(raster_paths, meta = "all", format = "data.table")
}
\arguments{
\item{raster_paths}{A vector of raster paths.}

\item{meta}{Vector with the desired metadata: one or more values among
'res', 'size', 'bbox', 'proj', 'unit', 'outformat', 'type'.
Alternatively meta = 'all' (default) allows to return all metadata.}

\item{format}{One between \code{data.table} (default), \code{data.frame} and \code{list}.}
}
\value{
A data.table, data.frame or list of the output metadata.
}
\description{
This accessory function extract some useful metadata from
a vector of raster paths.
}
\note{
License: GPL 3.0
}
\examples{
# Define product names
examplenames <- c(
  system.file("tif/L7_ETMs.tif", package="stars"),
  system.file("nc/bcsd_obs_1999.nc", package = "stars"),
  system.file("extdata/example_files/out_ref/S2A2A_20170703_022_Barbellino_BOA_10.tif", 
    package = "sen2r")
)

# Return metadata as data.table
raster_metadata(examplenames)

# Return metadata as list
raster_metadata(examplenames, format = "list")

# Return only two metadata
raster_metadata(examplenames, c("res", "unit"))

# Output with an invalid raster
examplenames <- c(examplenames, system.file("extdata/gdal_formats.json", package="sen2r"))
raster_metadata(examplenames)
}
\author{
Luigi Ranghetti, phD (2019) \email{ranghetti.l@irea.cnr.it}
}
