% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_download.R
\name{s2_download}
\alias{s2_download}
\title{Download S2 products.}
\usage{
s2_download(s2_prodlist = NULL, downloader = "builtin", apihub = NA,
  tile = NULL, outdir = ".")
}
\arguments{
\item{s2_prodlist}{List of the products to be downloaded
(this must be the output of \link{s2_list} function).}

\item{downloader}{Executable to use to download products
(default: "builtin"). Alternatives are "builtin" or "aria2"
(this requires aria2c to be installed).}

\item{apihub}{Path of the "apihub.txt" file containing credentials
of scihub account.
If NA (default), the default location inside the package will be used.}

\item{tile}{Single Sentinel-2 Tile string (5-length character)}

\item{outdir}{(optional) Full name of the existing output directory
where the files should be created (default: current directory).}

\item{overwrite_safe}{\code{logical} if FALSE, and a SAFE that should be downloaded is
already present in outdir, download is skipped, Default: FALSE}
}
\description{
The function downloads a single S2 product.
Input filename must be an element obtained with
\link{s2_list} function
(the content must be a URL, and the name the product name).
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
