% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_download.R
\name{s2_download}
\alias{s2_download}
\title{Download S2 products.}
\usage{
s2_download(filename = NULL, downloader = "wget", apihub = NULL,
  tile = NULL, write_dir = ".")
}
\arguments{
\item{filename}{TODO}

\item{downloader}{TODO}

\item{apihub}{TODO}

\item{tile}{TODO}

\item{write_dir}{TODO}
}
\value{
A vector of available products (being each element an URL,
and its name the product name)
}
\description{
The function downloads a single S2 product.
Input filename must be an element obtained with s2_list() function
(the content must be a URL, and the name the product name).
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
single_s2 <- "https://scihub.copernicus.eu/apihub/odata/v1/Products('c7142722-42bf-4f93-b8c5-59fd1792c430')/\\\\$value"
names(single_s2) <- "S2A_MSIL1C_20170613T101031_N0205_R022_T32TQQ_20170613T101608.SAFE"
# (this is equivalent to:
# single_s2 <- example_s2_list[1]
# where example_s2_list is the output of the example of the
# s2_list() function)

# Download the whole product
s2_download(single_s2, write_dir=tempdir()')

# Download a specific tile
s2_download(single_s2, tile="32TQQ", write_dir=tempdir()')
# (for products with compact names, the two produce equivalent
# results, while the forst downloads a SAFE archive, the second
# downloades single product files)
}

}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
