% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_getMetadata.R
\name{s2_getMetadata}
\alias{s2_getMetadata}
\title{Get information from S2 file name or metadata}
\usage{
s2_getMetadata(s2, info = "all")
}
\arguments{
\item{s2}{'character' or 'osgeo.gdal.Dataset' This input parameter
can be the main path of a S2 file, the path of the xml with metadata,
th path of a single granule, the xml path of a single granule, or a
'osgeo.gdal.Dataset' object (obtained reading the product with python).}

\item{info}{'character' (optional) Vector with the list of the metadata
which should be provided.
Accepted values are:
\itemize{
\item "all" (default): all the retrevable metadata are provided;
\item "fileinfo": only the metadata obtained by scanning the file name
and product structure (without opening it woth GDAL) are provided.
\item "nameinfo": only the metadata obtained by scanning the file name
are provided (it is faster and there is no need to have downloaded
yet the file; it works only with product names or granule names,
not with XML names).
\item a vector of single specific information (one or more from the
followings):
\itemize{
\item "prod_type" ('singlegranule' or 'product');
\item "version" ('old' or 'compact');
\item "tiles" (vector with the tiles ID available in the product);
\item "utm" (vector with the UTM zones used in the product);
\item "xml_main" (name of the main XML file with metadata);
\item "xml_granules" (names of the XML with granule metadata);
\item "level" ('1C' or '2A');
\item "creation_datetime", "id_tile", "mission", "centre",
"file_class", "id_orbit", "orbit_number", "sensing_datetime",
"id_baseline": metadata spefici of the product type and
version (they are returned only if obtainable for the specified
input);
\item "clouds","direction","orbit_n","preview_url", "proc_baseline",
"level", "sensing_datetime", "nodata_value", "saturated_value":
information retrieved from the metadata stored in the XML file.
}
}}
}
\value{
'list' list of the output metadata.
}
\description{
The function scan a Sentinel2 product (main path, granule path,
main / granule xml file or GDAL object) to retrieve information about
the product.
}
\note{
License: GPL 3.0
}
\examples{
# Define product name
s2_examplename <-
  "/path/of/the/product/S2A_MSIL1C_20170603T101031_N0205_R022_T32TQQ_20170603T101026.SAFE"

# Return only the information retrevable from the file names (files are not scanned)
s2_getMetadata(s2_examplename, info="nameinfo")

\dontrun{

# Return all the available information
s2_getMetadata(s2_examplename)

# Return some specific information
s2_getMetadata(s2_examplename, info=c("tiles", "level", "id_tile"))

# Return a single information
# (in this case, the output is a vector instead than a list)
s2_getMetadata(s2_examplename, info="level")

}
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
