% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_gui.R
\name{s2_gui}
\alias{s2_gui}
\title{Launch the GUI for Sentinel-2 products}
\usage{
s2_gui(param_list = NULL, thunderforest_api = NA)
}
\arguments{
\item{param_list}{List of parameters for initialising the GUI values
(if empty, default values are used).}

\item{thunderforest_api}{Character value with the API for thunderforest
layers (now not used).}
}
\value{
A list of parameters.
}
\description{
Launch the GUI to set parameters for the processing
chain of Sentinel-2 products.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
