% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_list.R
\name{s2_list}
\alias{s2_list}
\title{Retrieve list of available products.}
\usage{
s2_list(spatial_extent = NULL, tile = NULL, orbit = NULL,
  time_interval = NULL, corr_type = "auto", apihub = NULL,
  max_cloud = 110)
}
\arguments{
\item{spatial_extent}{TODO}

\item{tile}{TODO}

\item{orbit}{TODO}

\item{time_interval}{TODO}

\item{corr_type}{TODO}

\item{apihub}{TODO}

\item{max_cloud}{TODO}
}
\value{
A vector of available products (being each element an URL,
and its name the product name)
}
\description{
The function retrieves the list of available Sentinel-2
products basing on search criteria. It makes use of s2downoad
python function only to retrieve the list of files, without
downloading and correcting them.
}
\note{
License: GPL 3.0
}
\examples{
pos <- sp::SpatialPoints(data.frame("x"=12.0,"y"=44.8), proj4string=sp::CRS("+init=epsg:4326"))
time_window <- as.Date(c("2017-05-01","2017-07-30"))
example_s2_list <- s2_list(spatial_extent=pos, tile="32TQQ", time_interval=time_window)
print(example_s2_list)
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
