% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_mask.R
\name{s2_mask}
\alias{s2_mask}
\alias{s2_perc_masked}
\title{Apply cloud masks}
\usage{
s2_mask(infiles, maskfiles, mask_type = "cloud_medium_proba", smooth = 250,
  buffer = 250, max_mask = 80, outdir = "./masked", tmpdir = NA,
  rmtmp = TRUE, format = NA, subdirs = NA, compress = "DEFLATE",
  parallel = FALSE, overwrite = FALSE, .logfile_message = NA,
  .log_output = NA)

s2_perc_masked(infiles, maskfiles, mask_type = "cloud_medium_proba",
  tmpdir = NA, rmtmp = TRUE, parallel = FALSE)
}
\arguments{
\item{infiles}{A vector of input filenames. Input files are paths
of products already converted from SAFE format to a
format managed by GDAL (use \link{s2_translate} to do it);
their names must be in the sen2r naming convention
(\link{safe_shortname}).}

\item{maskfiles}{A vector of filenames from which to take the
information about cloud coverage (for now, only SCL products
have been implemented). It is not necessary that \code{maskfiles}
elements strictly match \code{infiles} ones. Input files are paths
of products already converted from SAFE format to a
format managed by GDAL (use \link{s2_translate} to do it);
their names must be in the sen2r naming convention
(\link{safe_shortname}).}

\item{mask_type}{(optional) Character vector which determines the type of
mask to be applied. Accepted values are:
\itemize{
\item "nodata": mask pixels checked as "No data" in the SCL product;
\item "cloud_high_proba": mask pixels checked as "No data" or
"Cloud (high probability)" in the SCL product;
\item "cloud_medium_proba": mask pixels checked as "No data" or
"Cloud (high or medium probability)" in the SCL product;
\item "cloud_low_proba": mask pixels checked as "No data" or
"Cloud (any probability)" in the SCL product;
\item "cloud_and_shadow": mask pixels checked as "No data",
"Cloud (any probability)" or "Cloud shadow" in the SCL product;
\item "cloud_shadow_cirrus": mask pixels checked as "No data",
"Cloud (any probability)", "Cloud shadow" or "Thin cirrus"
in the SCL product;
\item a string in the following form: "scl_n_m_n", where n, m and n are one or
more SCL class numbers (e.g. "scl_0_8_9_11"): mask pixels corresponding
to the classes specified in the string. E.g. string "scl_0_8_9_11" can
be used to mask classes 0 ("No data"), 8-9 ("Cloud (high or medium
probability)") and 11 ("Snow");
\item "opaque_clouds" (still to be implemented).
}}

\item{smooth}{(optional) Numerical (positive): should the mask be smoothed=the size (in the unit of
\code{inmask}, typically metres) to be used as radius for the smoothing
(the higher it is, the more smooth the output mask will result).}

\item{buffer}{(optional) Numerical (positive or negative): the size of the
buffer (in the unit of \code{inmask}, typically metres) to be applied to the
masked area after smoothing it (positive to enlarge, negative to reduce).}

\item{max_mask}{(optional) Numeric value (range 0 to 100), which represents
the maximum percentage of allowed masked surface (by clouds or any other
type of mask chosen with argument \code{mask_type}) for producing outputs.
Images with a percentage of masked surface greater than \code{max_mask}%
are not processed (the list of expected output files which have not been
generated is returned as an attribute, named "skipped").
Default value is 80.
Notice that the percentage is computed on non-NA values (if input images
had previously been clipped and masked using a polygon, the percentage is
computed on the surface included in the masking polygons).}

\item{outdir}{(optional) Full name of the output directory where
the files should be created (default: "current directory"masked"
subdir of current directory).
\code{outdir} can bot be an existing or non-existing directory (in the
second case, its parent directory must exists).
If it is a relative path, it is expanded from the common parent
directory of \code{infiles}.}

\item{tmpdir}{(optional) Path where intermediate files (VRT) will be created.
Default is a temporary directory.}

\item{rmtmp}{(optional) Logical: should temporary files be removed?
(Default: TRUE).
This parameter takes effect only if the output files are not VRT
(in this case temporary files cannot be deleted, because rasters of source
bands are included within them).}

\item{format}{(optional) Format of the output file (in a
format recognised by GDAL). Default is the same format of input images
(or "GTiff" in case of VRT input images).}

\item{subdirs}{(optional) Logical: if TRUE, different indices are
placed in separated \code{outfile} subdirectories; if FALSE, they are placed in
\code{outfile} directory; if NA (default), subdirectories are created only if
more than a single spectral index is required.}

\item{compress}{(optional) In the case a GTiff format is
present, the compression indicated with this parameter is used.}

\item{parallel}{(optional) Logical: if TRUE, masking is conducted using parallel
processing, to speed-up the computation for large rasters.
The number of cores is automatically determined; specifying it is also
possible (e.g. \code{parallel = 4}).
If FALSE (default), single core processing is used.
Multiprocess masking computation is always performed in singlecore mode
if \code{format != "VRT"} (because in this case there is no gain in using
multicore processing).}

\item{overwrite}{(optional) Logical value: should existing output files be
overwritten? (default: FALSE)}

\item{.logfile_message}{(optional) Internal parameter
(it is used when the function is called by \code{sen2r()}).}

\item{.log_output}{(optional) Internal parameter
(it is used when the function is called by \code{sen2r()}).}
}
\value{
\link{s2_mask} returns a vector with the names of the created products.
An attribute "toomasked" contains the paths of the outputs which were not
created cause to the high percentage of cloud coverage.

\link{s2_perc_masked} returns a names vector with the percentages
of masked surtfaces.
}
\description{
\link{s2_mask} Applies a cloud mask to a Sentinel-2 product. Since
\link{raster} functions are used to perform computations, output files
are physical rasters (no output VRT is allowed).

\link{s2_perc_masked} computes the percentage of cloud-masked surface.
The function is similar to \link{s2_mask}, but it returns percentages instead
of masked rasters.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
