% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_order.R
\name{s2_order}
\alias{s2_order}
\title{Order S2 products.}
\usage{
s2_order(s2_prodlist = NULL, apihub = NA)
}
\arguments{
\item{s2_prodlist}{List of the products to be ordered
(this must be the output of \link{s2_list} function).}

\item{apihub}{Path of the "apihub.txt" file containing credentials
of SciHub account.
If NA (default), the default location inside the package will be used.}
}
\value{
A named vector, containing the selection of \code{s2_prodlist} elements
which were ordered.
Moreover, the vector includes two attributes:
\itemize{
\item "available" with the elements of \code{s2_prodlist} which were already
available for download,
\item "notordered" with the elements of \code{s2_prodlist} which were not ordered
for any reasons.
}
}
\description{
The function orders S2 products from Long Term Archive
(https://scihub.copernicus.eu/userguide/LongTermArchive).
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
single_s2 <- paste0("https://scihub.copernicus.eu/apihub/odata/v1/",
  "Products(\\'c7142722-42bf-4f93-b8c5-59fd1792c430\\')/$value")
names(single_s2) <- "S2A_MSIL1C_20170613T101031_N0205_R022_T32TQQ_20170613T101608.SAFE"
# (this is equivalent to:
# single_s2 <- example_s2_list[1]
# where example_s2_list is the output of the example of the
# s2_list() function)

# Order the product
ordered_prods <- s2_order(single_s2)
}

# Check in a second time if the product was made available
safe_is_online(ordered_prods)
}
\author{
Luigi Ranghetti, phD (2019) \email{luigi@ranghetti.info}
}
