% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_order.R
\name{s2_order}
\alias{s2_order}
\title{Order S2 products (deprecated).}
\usage{
s2_order(
  s2_prodlist = NULL,
  export_prodlist = TRUE,
  delay = 0.5,
  apihub = NA,
  service = NA,
  reorder = TRUE
)
}
\arguments{
\item{s2_prodlist}{deprecated}

\item{export_prodlist}{deprecated}

\item{delay}{deprecated}

\item{apihub}{deprecated}

\item{service}{deprecated}

\item{reorder}{deprecated}
}
\value{
deprecated
}
\description{
This function is deprecated and will be removed.
}
