% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_shortname.R
\name{s2_shortname}
\alias{s2_shortname}
\title{Rename products using a shorten convention}
\usage{
s2_shortname(prod_name)
}
\arguments{
\item{input_name}{Input Sentinel-2 product name: it can be both
a complete path or a simple file name. If the file exists, the product
is scanned to retrieve the tiles list and UTM zones; otherwise the
basename is simply translated in he short naming convention.}
}
\value{
Output product name
}
\description{
This function renames a Sentinel-2 product in order to
obtain shorten names. See the details for the structure of the
adopted schema. The function applies only to product names (not
to single granule names), since it is thought to be applied to
entire products.
}
\details{
\href{https://earth.esa.int/web/sentinel/user-guides/sentinel-2-msi/naming-convention}{ESA Sentinel-2 naming convention}
is particularly long-winded; moreover, the convention changed after
December 6th 2016, causing products to follow two different schemes.

The convention here adopted follows this schema:

\code{S2mll_yyyymmdd_rrr_ttttt}

where:
\itemize{
\item \code{S2mll} shows the mission ID (\code{S2A} or \code{S2B}) and the product level
(\code{1C} or \code{2A});
\item \code{yyyymmdd} is the sensing date (e.g. \code{20170603} for 2017-06-03) ; the
hour is skipped, since a single sensor can not pass two times in a
day on the same tile);
\item \code{rrr} is the relative orbit number (e.g. \code{022});
\item \code{ttttt} is the tile number (e.g. \code{32TQQ}).
}

Products which contains only one tile (date ≥ 2016-12-06 or
< 2016-12-06 but containing only one tile) follows entirely this convention;
for products with more than one tile (or that can not be checked), the field
\code{ttttt} is replaced by:
\itemize{
\item \code{uuxxx} if the UTM zone is unique (where \code{uu} is the UTM zone, and \code{xxx}
is a random numeric sequence), e.g. \code{32876};
\item \code{xxxxx} if the UTM zone is not unique or it it can not be checked (being
\code{xxxxx} a random literal A-Z sequence).
}

Since the tile number is composed by two numeric and three literal character,
these two conventions (one composed by five numeric and the other by five
literal characters) ensure that these three cases can be distinguished.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
