% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_thumbnails.R
\name{s2_thumbnails}
\alias{s2_thumbnails}
\title{Create thumbnails for S2 products.}
\usage{
s2_thumbnails(infiles, prod_type = NA, rgb_type = "SwirNirR", dim = 1024,
  scaleRange = NA, outdir = NA, tmpdir = NA, overwrite = FALSE)
}
\arguments{
\item{infiles}{A vector of input filenames. Input files are paths
of products already converted from SAFE format to a
format managed by GDAL (use \link{s2_translate} to do it);
their names must be in the SALTO-S2 naming convention
(\link{s2_shortname}).}

\item{prod_type}{(optional) Output product (see \link{s2_shortname} for the
list of accepted products). If not provided, it is retrieved from the
file name.}

\item{rgb_type}{(optional) For BOA and TOA products, this value determine
the type of false colours to be used for the thumbnails:
\itemize{
\item \code{"SwirNirR"} (default) for SWIR-NIR-Red;
\item \code{"NirRG"} for NIR-Red-Green;
\item \code{"RGB"} for true colours;
}}

\item{dim}{Integer value, with the maximum greater dimension in pixels (width or
height) of the output images (default: 1024 px).
If this is lower than the corresponding dimension of the maps, maps are
rescaled before producing the thumbnails; otherwise the original dimensions
are maintained.
To keep the original size in any case, set \code{dim = Inf}.}

\item{scaleRange}{(optional) Range of valid values. If not specified, it is
automatically retrieved from the product type. It is useful for BOA "dark"
products. When spectral indices are saved with Integer values, the range
is automatically set to -10000 to 10000.}

\item{outdir}{(optional) Full name of the existing output directory
where the files should be created.  Default is a subdirectory (named
"thumbnails") of the parent directory of each input file.}

\item{tmpdir}{(optional) Path where intermediate VRT will be created.
Default is a temporary directory.}

\item{overwrite}{(optional) Logical value: should existing thumbnails be
overwritten? (default: TRUE)}
}
\value{
A vector with the names of the created images.
}
\description{
Function to create thumbnail images for Sentinel-2
products. BOA and TOA multiband images are rendered as false colour
JPEG images; SCL maps are rendered as 8-bit PNG;
other singleband images (like spectral indices) are rendered as
JPEG images with a standard colour palette.
Some improvements still have to be done:
\itemize{
\item allowing the possibility to chose the palette and the limits
(for now, the range -1 to 1 is always used);
\item adding support for RGB products.
}

Output images are georeferenced.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
