% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_is_online.R
\name{safe_is_online}
\alias{safe_is_online}
\title{Check if SAFE is available for download (deprecated)}
\usage{
safe_is_online(s2_prodlist = NULL, apihub = NA, verbose = TRUE)
}
\arguments{
\item{s2_prodlist}{deprecated}

\item{apihub}{deprecated}

\item{verbose}{deprecated}
}
\value{
deprecated
}
\description{
This function is deprecated and will be removed.
}
