% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_thumbnails.R
\name{stack2rgb}
\alias{stack2rgb}
\title{Produce an RGB image from a multiband raster file.}
\usage{
stack2rgb(in_rast, out_file = NULL, bands = 1:3, minval = 0,
  maxval = 10000, tmpdir = NA)
}
\arguments{
\item{in_rast}{Path of the input multiband raster.}

\item{out_file}{(optional) Path of the output RGB JPEG image; if NULL
(default), a RasterBrick will be returned.}

\item{bands}{(optional) 3-length integer argument, with the position of
the three bands to be used respectively for red, green and blue.}

\item{minval}{(optional) the value corresponding to black (default: 0).}

\item{maxval}{(optional) the value corresponding to white (default: 10000).}

\item{tmpdir}{(optional) Path where intermediate files will be created.
Default is a temporary directory.}
}
\value{
The path of the output image; alternatively, the output image
as RasterBrick (if \code{out_rast = NULL}).
}
\description{
Internal function to create JPEG images from a multiband raster
file. This function is used by \link{s2_thumbnails}, and it will be exported
when it would be more generalised.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
